@@ -86,68 +120,320 @@
 	"author": "Bitfocus AS",
 	"license": "MIT",
 	"devDependencies": {
+		"asar": "^3.0.3",
+		"aws-sdk": "^2.783.0",
 		"chai": "^4.2.0",
 		"chai-fs": "^2.0.0",
-		"electron": "^2.0.8",
-		"electron-builder": "20.10.0",
-		"mocha": "^6.1.4",
-		"nodemon": "^2.0.2",
-		"s3": "^4.4.0",
+		"cross-env": "^7.0.3",
+		"electron": "^11.3.0",
+		"electron-builder": "^22.10.5",
+		"mocha": "^6.2.3",
+		"nodemon": "^2.0.7",
+		"p-all": "^3.0.0",
+		"prettier": "^2.2.1",
+		"rimraf": "^3.0.2",
 		"s3-upload-stream": "^1.0.7"
 	},
 	"dependencies": {
-		"@coreui/coreui": "^2.0.0-rc.2",
-		"@coreui/coreui-plugin-chartjs-custom-tooltips": "1.2.0",
+		"@sentry/electron": "^1.3.0",
 		"app-root-path": "^2.0.1",
-		"aws-sdk": "^2.656.0",
 		"binopsy": "^0.0.0",
-		"bootstrap": "^4.2.1",
-		"bootstrap-colorpicker": "^2.5.2",
-		"chart.js": "^2.7.2",
 		"check-ip": "^1.1.1",
-		"debug": "^4.1.1",
-		"electron-debug": "^1.5.0",
-		"electron-rpc": "^2.0.1",
-		"elgato-stream-deck": "^3.3.2",
+		"companion-module-7thsensedesign-delta": "github:bitfocus/companion-module-7thsensedesign-delta#v1.0.1",
+		"companion-module-agf-characterworks": "github:bitfocus/companion-module-agf-characterworks#v1.0.3",
+		"companion-module-aja-helo": "github:bitfocus/companion-module-aja-helo#v1.0.2",
+		"companion-module-aja-kipro": "github:bitfocus/companion-module-aja-kipro#v2.0.9",
+		"companion-module-aja-kumo": "github:bitfocus/companion-module-aja-kumo#v1.0.2",
+		"companion-module-allenheath-dlive-ilive": "github:bitfocus/companion-module-allenheath-dlive-ilive#v1.3.5",
+		"companion-module-allenheath-qu": "github:bitfocus/companion-module-allenheath-qu#v1.0.6",
+		"companion-module-allenheath-sq": "github:bitfocus/companion-module-allenheath-sq#v1.3.4",
+		"companion-module-analogway-eks500": "github:bitfocus/companion-module-analogway-eks500#v1.0.1",
+		"companion-module-analogway-livecore": "github:bitfocus/companion-module-analogway-livecore#v1.1.1",
+		"companion-module-analogway-livepremier": "github:bitfocus/companion-module-analogway-livepremier#v1.0.2",
+		"companion-module-analogway-midra": "github:bitfocus/companion-module-analogway-midra#v1.0.2",
+		"companion-module-analogway-picturall": "github:bitfocus/companion-module-analogway-picturall#v1.2.1",
+		"companion-module-analogway-pls300": "github:bitfocus/companion-module-analogway-pls300#v1.0.3",
+		"companion-module-analogway-vertige": "github:bitfocus/companion-module-analogway-vertige#v1.0.0",
+		"companion-module-analogway-vio": "github:bitfocus/companion-module-analogway-vio#v1.0.1",
+		"companion-module-anomes-millumin": "github:bitfocus/companion-module-anomes-millumin#v1.0.6",
+		"companion-module-aten-matrix": "github:bitfocus/companion-module-aten-matrix#v1.0.0",
+		"companion-module-audiostrom-liveprofessor": "github:bitfocus/companion-module-audiostrom-liveprofessor#v1.0.0",
+		"companion-module-audivero-unityintercom-client": "github:bitfocus/companion-module-audivero-unityintercom-client#v1.0.1",
+		"companion-module-avishop-hdbaset-matrix": "github:bitfocus/companion-module-avishop-hdbaset-matrix#v1.0.1",
+		"companion-module-avolites-ai": "github:bitfocus/companion-module-avolites-ai#v1.0.1",
+		"companion-module-avolites-titan": "github:bitfocus/companion-module-avolites-titan#v1.1.0",
+		"companion-module-avproconnect-acmx1616-auhd": "github:bitfocus/companion-module-avproconnect-acmx1616-auhd#v1.0.0",
+		"companion-module-avstumpfl-pixera": "github:bitfocus/companion-module-avstumpfl-pixera#v1.0.2",
+		"companion-module-barco-dcs": "github:bitfocus/companion-module-barco-dcs#v1.0.4",
+		"companion-module-barco-dp": "github:bitfocus/companion-module-barco-dp#v1.1.2",
+		"companion-module-barco-encore": "github:bitfocus/companion-module-barco-encore#v1.0.1",
+		"companion-module-barco-eventmaster": "github:bitfocus/companion-module-barco-eventmaster#v1.3.5",
+		"companion-module-barco-eventmaster-xml": "github:bitfocus/companion-module-barco-eventmaster-xml#v1.2.4",
+		"companion-module-barco-hdx": "github:bitfocus/companion-module-barco-hdx#v1.1.5",
+		"companion-module-barco-imagepro": "github:bitfocus/companion-module-barco-imagepro#v1.0.1",
+		"companion-module-barco-matrixpro": "github:bitfocus/companion-module-barco-matrixpro#v1.1.0",
+		"companion-module-barco-pds": "github:bitfocus/companion-module-barco-pds#v1.1.5",
+		"companion-module-barco-pulse": "github:bitfocus/companion-module-barco-pulse#1.1.2",
+		"companion-module-bbc-raven": "github:bitfocus/companion-module-bbc-raven#v1.0.0",
+		"companion-module-behringer-wing": "github:bitfocus/companion-module-behringer-wing#v1.0.3",
+		"companion-module-behringer-x32": "github:bitfocus/companion-module-behringer-x32#v2.4.3",
+		"companion-module-behringer-xair": "github:bitfocus/companion-module-behringer-xair#v1.6.2",
+		"companion-module-birddog-studio": "github:bitfocus/companion-module-birddog-studio#v1.0.0",
+		"companion-module-birddog-visca": "github:bitfocus/companion-module-birddog-visca#v1.0.3",
+		"companion-module-bitfocus-companion": "github:bitfocus/companion-module-bitfocus-companion#v1.1.6",
+		"companion-module-blackbox-boxilla": "github:bitfocus/companion-module-blackbox-boxilla#v1.0.2",
+		"companion-module-blackdiamondvideo-phantom800": "github:bitfocus/companion-module-blackdiamondvideo-phantom800#v1.0.0",
+		"companion-module-bmd-atem": "github:bitfocus/companion-module-bmd-atem#v2.9.3",
+		"companion-module-bmd-hyperdeck": "github:bitfocus/companion-module-bmd-hyperdeck#v1.2.1",
+		"companion-module-bmd-multiview16": "github:bitfocus/companion-module-bmd-multiview16#v1.0.1",
+		"companion-module-bmd-multiview4": "github:bitfocus/companion-module-bmd-multiview4#v1.3.2",
+		"companion-module-bmd-smartview": "github:bitfocus/companion-module-bmd-smartview#v1.1.2",
+		"companion-module-bmd-teranex": "github:bitfocus/companion-module-bmd-teranex#v1.0.4",
+		"companion-module-bmd-videohub": "github:bitfocus/companion-module-bmd-videohub#v1.3.0",
+		"companion-module-borealsystems-director": "github:bitfocus/companion-module-borealsystems-director#v1.0.0",
+		"companion-module-brightsign-player": "github:bitfocus/companion-module-brightsign-player#v1.1.1",
+		"companion-module-casparcg-server": "github:bitfocus/companion-module-casparcg-server#v1.0.2",
+		"companion-module-chamsys-magicq-osc": "github:bitfocus/companion-module-chamsys-magicq-osc#v1.0.1",
+		"companion-module-chamsys-magicq-udp": "github:bitfocus/companion-module-chamsys-magicq-udp#v1.0.2",
+		"companion-module-christie-pandorasbox": "github:bitfocus/companion-module-christie-pandorasbox#v2.0.2",
+		"companion-module-christie-projector": "github:bitfocus/companion-module-christie-projector#v2.0.3",
+		"companion-module-christie-spyder": "github:bitfocus/companion-module-christie-spyder#v1.0.0",
+		"companion-module-christie-wd": "github:bitfocus/companion-module-christie-wd#v1.0.6",
+		"companion-module-cisco-cms": "github:bitfocus/companion-module-cisco-cms#v1.0.1",
+		"companion-module-cockos-reaper": "github:bitfocus/companion-module-cockos-reaper#v1.1.0",
+		"companion-module-crystal-scte": "github:bitfocus/companion-module-crystal-scte#v1.0.0",
+		"companion-module-dahuasecurity-ptz": "github:bitfocus/companion-module-dahuasecurity-ptz#v1.0.1",
+		"companion-module-dalite-scb": "github:bitfocus/companion-module-dalite-scb#v1.0.0",
+		"companion-module-dashare-multiplay": "github:bitfocus/companion-module-dashare-multiplay#v1.0.1",
+		"companion-module-datapath-fx4": "github:bitfocus/companion-module-datapath-fx4#v1.0.0",
+		"companion-module-dataton-watchout": "github:bitfocus/companion-module-dataton-watchout#v1.2.0",
+		"companion-module-datavideo-dvip": "github:bitfocus/companion-module-datavideo-dvip#v1.0.0",
+		"companion-module-datavideo-visca": "github:bitfocus/companion-module-datavideo-visca#v1.0.3",
+		"companion-module-denon-dn-500bd-mkii": "github:bitfocus/companion-module-denon-dn-500bd-mkii#v1.0.2",
+		"companion-module-denon-receiver": "github:bitfocus/companion-module-denon-receiver#v1.0.1",
+		"companion-module-denon-recorder": "github:bitfocus/companion-module-denon-recorder#v1.0.1",
+		"companion-module-depili-clock-8001": "github:bitfocus/companion-module-depili-clock-8001#v5.1.1",
+		"companion-module-dexon-dimax": "github:bitfocus/companion-module-dexon-dimax#v1.0.3",
+		"companion-module-dexon-divip": "github:bitfocus/companion-module-dexon-divip#v1.0.2",
+		"companion-module-dexon-matrix": "github:bitfocus/companion-module-dexon-matrix#v1.0.1",
+		"companion-module-digico-osc": "github:bitfocus/companion-module-digico-osc#v1.0.2",
+		"companion-module-digitalprojection-highlight": "github:bitfocus/companion-module-digitalprojection-highlight#v1.0.2",
+		"companion-module-disguise": "github:bitfocus/companion-module-disguise#v1.1.0",
+		"companion-module-disguise-mtc": "github:bitfocus/companion-module-disguise-mtc#v1.0.2",
+		"companion-module-dolby-cinemaprocessor": "github:bitfocus/companion-module-dolby-cinemaprocessor#v1.0.0",
+		"companion-module-draco-tera": "github:bitfocus/companion-module-draco-tera#v1.0.4",
+		"companion-module-dsan-limitimer": "github:bitfocus/companion-module-dsan-limitimer#v1.1.1",
+		"companion-module-dsan-perfectcue": "github:bitfocus/companion-module-dsan-perfectcue#v1.0.0",
+		"companion-module-dtvideolabs-playbackproplus": "github:bitfocus/companion-module-dtvideolabs-playbackproplus#v1.0.2",
+		"companion-module-elgato-keylight": "github:bitfocus/companion-module-elgato-keylight#v1.2.0",
+		"companion-module-epiphan-pearl": "github:bitfocus/companion-module-epiphan-pearl#v1.0.6",
+		"companion-module-etc-eos": "github:bitfocus/companion-module-etc-eos#v1.2.1",
+		"companion-module-extron-dxp": "github:bitfocus/companion-module-extron-dxp#v1.0.8",
+		"companion-module-extron-in1604": "github:bitfocus/companion-module-extron-in1604#v1.0.3",
+		"companion-module-extron-ipl-t-pcs4": "github:bitfocus/companion-module-extron-ipl-t-pcs4#v1.0.2",
+		"companion-module-extron-smp111": "github:bitfocus/companion-module-extron-smp111#v1.0.4",
+		"companion-module-extron-smp351": "github:bitfocus/companion-module-extron-smp351#v1.1.3",
+		"companion-module-extron-smx": "github:bitfocus/companion-module-extron-smx#v1.0.4",
+		"companion-module-faithchapel-videoplayoutserver": "github:bitfocus/companion-module-faithchapel-videoplayoutserver#v1.0.1",
+		"companion-module-figure53-go-button": "github:bitfocus/companion-module-figure53-go-button#v1.2.3",
+		"companion-module-figure53-qlab": "github:bitfocus/companion-module-figure53-qlab#v1.2.4",
+		"companion-module-figure53-qlab-advance": "github:bitfocus/companion-module-figure53-qlab-advance#v1.3.5",
+		"companion-module-figure53-qview": "github:bitfocus/companion-module-figure53-qview#v1.0.0",
+		"companion-module-folivora-btt": "github:bitfocus/companion-module-folivora-btt#v1.0.2",
+		"companion-module-fora-hvs": "github:bitfocus/companion-module-fora-hvs#v1.2.3",
+		"companion-module-foscam-ptz": "github:bitfocus/companion-module-foscam-ptz#v1.0.1",
+		"companion-module-gallery-virtualvtrpro": "github:bitfocus/companion-module-gallery-virtualvtrpro#v1.0.1",
+		"companion-module-gammacontrol-gmaestro": "github:bitfocus/companion-module-gammacontrol-gmaestro#v1.0.2",
+		"companion-module-gefen-dvimatrix": "github:bitfocus/companion-module-gefen-dvimatrix#v1.0.1",
+		"companion-module-generic-artnet": "github:bitfocus/companion-module-generic-artnet#v1.0.2",
+		"companion-module-generic-emberplus": "github:bitfocus/companion-module-generic-emberplus#v1.0.1",
+		"companion-module-generic-http": "github:bitfocus/companion-module-generic-http#v1.0.8",
+		"companion-module-generic-mqtt": "github:bitfocus/companion-module-generic-mqtt#v1.1.0",
+		"companion-module-generic-osc": "github:bitfocus/companion-module-generic-osc#v1.0.2",
+		"companion-module-generic-pjlink": "github:bitfocus/companion-module-generic-pjlink#v1.0.5",
+		"companion-module-generic-sacn": "github:bitfocus/companion-module-generic-sacn#v1.0.0",
+		"companion-module-generic-tcp-serial": "github:bitfocus/companion-module-generic-tcp-serial#v1.0.2",
+		"companion-module-generic-tcp-udp": "github:bitfocus/companion-module-generic-tcp-udp#v1.0.6",
+		"companion-module-generic-wakeonlan": "github:bitfocus/companion-module-generic-wakeonlan#v1.0.4",
+		"companion-module-globalcache-itac-cc": "github:bitfocus/companion-module-globalcache-itac-cc#v1.0.1",
+		"companion-module-globalcache-itac-ir": "github:bitfocus/companion-module-globalcache-itac-ir#v1.0.5",
+		"companion-module-globalcache-itac-sl": "github:bitfocus/companion-module-globalcache-itac-sl#v1.0.3",
+		"companion-module-grassvalley-amp": "github:bitfocus/companion-module-grassvalley-amp#v1.1.3",
+		"companion-module-greenhippo-hippotizer": "github:bitfocus/companion-module-greenhippo-hippotizer#v1.0.1",
+		"companion-module-h2r-graphics": "github:bitfocus/companion-module-h2r-graphics#v1.1.1",
+		"companion-module-haivision-connectdvr": "github:bitfocus/companion-module-haivision-connectdvr#v1.0.9",
+		"companion-module-highend-hog4": "github:bitfocus/companion-module-highend-hog4#v1.1.0",
+		"companion-module-hologfx-holographics": "github:bitfocus/companion-module-hologfx-holographics#v1.0.0",
+		"companion-module-homeassistant-server": "github:bitfocus/companion-module-homeassistant-server#v0.3.1",
+		"companion-module-ifelseware-avplayback": "github:bitfocus/companion-module-ifelseware-avplayback#v1.0.1",
+		"companion-module-imimot-mitti": "github:bitfocus/companion-module-imimot-mitti#v1.0.6",
+		"companion-module-interactivetechnologies-cueserver": "github:bitfocus/companion-module-interactivetechnologies-cueserver#v1.0.1",
+		"companion-module-irisdown-countdowntimer": "github:bitfocus/companion-module-irisdown-countdowntimer#v1.1.3",
+		"companion-module-irisdown-remoteshowcontrol": "github:bitfocus/companion-module-irisdown-remoteshowcontrol#v1.0.2",
+		"companion-module-jamesholt-x32tc": "github:bitfocus/companion-module-jamesholt-x32tc#v1.0.6",
+		"companion-module-joy-playdeck": "github:bitfocus/companion-module-joy-playdeck#v1.0.1",
+		"companion-module-justmacros-lua": "github:bitfocus/companion-module-justmacros-lua#v1.0.1",
+		"companion-module-jvc-ptz": "github:bitfocus/companion-module-jvc-ptz#v1.0.9",
+		"companion-module-kramer-matrix": "github:bitfocus/companion-module-kramer-matrix#v1.2.1",
+		"companion-module-kramer-vp727": "github:bitfocus/companion-module-kramer-vp727#v1.0.1",
+		"companion-module-kramer-vp734": "github:bitfocus/companion-module-kramer-vp734#v1.0.0",
+		"companion-module-kramer-vp773a": "github:bitfocus/companion-module-kramer-vp773a#v1.0.0",
+		"companion-module-kramer-vs41h": "github:bitfocus/companion-module-kramer-vs41h#v1.0.0",
+		"companion-module-lectrosonics-aspen": "github:bitfocus/companion-module-lectrosonics-aspen#v1.0.0",
+		"companion-module-lightware-lw2": "github:bitfocus/companion-module-lightware-lw2#v1.0.2",
+		"companion-module-lightware-lw3": "github:bitfocus/companion-module-lightware-lw3#v1.0.1",
+		"companion-module-liminalet-zoomosc": "github:bitfocus/companion-module-liminalet-zoomosc#v1.0.4",
+		"companion-module-linkbox-remote": "github:bitfocus/companion-module-linkbox-remote#v1.2.1",
+		"companion-module-livingasone-decoders": "github:bitfocus/companion-module-livingasone-decoders#v1.1.2",
+		"companion-module-lumens-visca": "github:bitfocus/companion-module-lumens-visca#v1.0.1",
+		"companion-module-lyntec-rpc-breaker": "github:bitfocus/companion-module-lyntec-rpc-breaker#v1.0.3",
+		"companion-module-magewell-proconvert-decoder": "github:bitfocus/companion-module-magewell-proconvert-decoder#v1.0.0",
+		"companion-module-magewell-ultrastream": "github:bitfocus/companion-module-magewell-ultrastream#v0.0.3",
+		"companion-module-magicsoft-recorder": "github:bitfocus/companion-module-magicsoft-recorder#v1.0.0",
+		"companion-module-malighting-grandma2": "github:bitfocus/companion-module-malighting-grandma2#v1.0.10",
+		"companion-module-malighting-msc": "github:bitfocus/companion-module-malighting-msc#v0.2.1",
+		"companion-module-matrox-monarch": "github:bitfocus/companion-module-matrox-monarch#v1.0.0",
+		"companion-module-media-player-classic": "github:bitfocus/companion-module-media-player-classic#v1.0.0",
+		"companion-module-metus-ingest": "github:bitfocus/companion-module-metus-ingest#v1.2.1",
+		"companion-module-middleatlantic-racklink": "github:bitfocus/companion-module-middleatlantic-racklink#v1.0.0",
+		"companion-module-modelighting-edin": "github:bitfocus/companion-module-modelighting-edin#v1.0.0",
+		"companion-module-modulo": "github:bitfocus/companion-module-modulo#v1.0.1",
+		"companion-module-motu-avb": "github:bitfocus/companion-module-motu-avb#v1.0.1",
+		"companion-module-msc-router": "github:bitfocus/companion-module-msc-router#v1.0.2",
+		"companion-module-multicamsystems-multicamsuite": "github:bitfocus/companion-module-multicamsystems-multicamsuite#v1.0.1",
+		"companion-module-neodarque-stagetimer2": "github:bitfocus/companion-module-neodarque-stagetimer2#v1.2.5",
+		"companion-module-netio-powerbox": "github:bitfocus/companion-module-netio-powerbox#v1.0.0",
+		"companion-module-nevion-mrp": "github:bitfocus/companion-module-nevion-mrp#v2.0.2",
+		"companion-module-newbluefx-titler": "github:bitfocus/companion-module-newbluefx-titler#v1.0.1",
+		"companion-module-newtek-ndistudiomonitor": "github:bitfocus/companion-module-newtek-ndistudiomonitor#v1.0.1",
+		"companion-module-newtek-tricaster": "github:bitfocus/companion-module-newtek-tricaster#v1.0.8",
+		"companion-module-nexo-nxamp": "github:bitfocus/companion-module-nexo-nxamp#v1.0.0",
+		"companion-module-noismada-octopuslistener": "github:bitfocus/companion-module-noismada-octopuslistener#v1.0.2",
+		"companion-module-noismada-octopusshowcontrol": "github:bitfocus/companion-module-noismada-octopusshowcontrol#v1.2.0",
+		"companion-module-novastar-controller": "github:bitfocus/companion-module-novastar-controller#v1.0.6",
+		"companion-module-obs-studio": "github:bitfocus/companion-module-obs-studio#v1.0.18",
+		"companion-module-obsidiancontrol-onyx": "github:bitfocus/companion-module-obsidiancontrol-onyx#v1.0.3",
+		"companion-module-octava-pro-dsx": "github:bitfocus/companion-module-octava-pro-dsx#v0.0.1",
+		"companion-module-openlp-http": "github:bitfocus/companion-module-openlp-http#v0.1.0",
+		"companion-module-opensong-api": "github:bitfocus/companion-module-opensong-api#v1.0.1",
+		"companion-module-openweather-rest": "github:bitfocus/companion-module-openweather-rest#v1.0.2",
+		"companion-module-opticis-omm-1000": "github:bitfocus/companion-module-opticis-omm-1000#v1.0.2",
+		"companion-module-optoma-z28s": "github:bitfocus/companion-module-optoma-z28s#v1.0.2",
+		"companion-module-orfast-ndi": "github:bitfocus/companion-module-orfast-ndi#v1.0.0",
+		"companion-module-panasonic-avhs": "github:bitfocus/companion-module-panasonic-avhs#v1.0.1",
+		"companion-module-panasonic-camera-controller": "github:bitfocus/companion-module-panasonic-camera-controller#v1.0.3",
+		"companion-module-panasonic-projector": "github:bitfocus/companion-module-panasonic-projector#v1.0.1",
+		"companion-module-panasonic-ptz": "github:bitfocus/companion-module-panasonic-ptz#v1.0.14",
+		"companion-module-panasonic-tv-th": "github:bitfocus/companion-module-panasonic-tv-th#v0.0.1",
+		"companion-module-pangolin-beyond": "github:bitfocus/companion-module-pangolin-beyond#v1.0.1",
+		"companion-module-phillips-hue": "github:bitfocus/companion-module-phillips-hue#v0.0.3",
+		"companion-module-pixap-pixtimerpro": "github:bitfocus/companion-module-pixap-pixtimerpro#v1.0.6",
+		"companion-module-planningcenter-serviceslive": "github:bitfocus/companion-module-planningcenter-serviceslive#v1.0.5",
+		"companion-module-presentationtools-aps": "github:bitfocus/companion-module-presentationtools-aps#v1.0.0",
+		"companion-module-presentationtools-cuetimer": "github:bitfocus/companion-module-presentationtools-cuetimer#v1.0.0",
+		"companion-module-protopie-bridge": "github:bitfocus/companion-module-protopie-bridge#v1.0.0",
+		"companion-module-prsi-ipower": "github:bitfocus/companion-module-prsi-ipower#v1.0.2",
+		"companion-module-ptzoptics-visca": "github:bitfocus/companion-module-ptzoptics-visca#v1.1.6",
+		"companion-module-qsys-remote-control": "github:bitfocus/companion-module-qsys-remote-control#v1.0.3",
+		"companion-module-radiodj-rest": "github:bitfocus/companion-module-radiodj-rest#v1.0.0",
+		"companion-module-rationalacoustics-smaart3": "github:bitfocus/companion-module-rationalacoustics-smaart3#v1.0.3",
+		"companion-module-renewedvision-propresenter": "github:bitfocus/companion-module-renewedvision-propresenter#v2.3.7",
+		"companion-module-renewedvision-pvp": "github:bitfocus/companion-module-renewedvision-pvp#v1.0.7",
+		"companion-module-resolume-arena": "github:bitfocus/companion-module-resolume-arena#v1.0.3",
+		"companion-module-roku-tv": "github:bitfocus/companion-module-roku-tv#v1.0.1",
+		"companion-module-roland-m5000": "github:bitfocus/companion-module-roland-m5000#v1.0.1",
+		"companion-module-roland-v1200hd": "github:bitfocus/companion-module-roland-v1200hd#v1.0.0",
+		"companion-module-roland-v600uhd": "github:bitfocus/companion-module-roland-v600uhd#v1.0.1",
+		"companion-module-roland-v60hd": "github:bitfocus/companion-module-roland-v60hd#v1.0.6",
+		"companion-module-roland-vp42h": "github:bitfocus/companion-module-roland-vp42h#v1.0.0",
+		"companion-module-roland-vr50hd-mk2": "github:bitfocus/companion-module-roland-vr50hd-mk2#v1.0.0",
+		"companion-module-roland-xs42h": "github:bitfocus/companion-module-roland-xs42h#v1.0.0",
+		"companion-module-roland-xs62s": "github:bitfocus/companion-module-roland-xs62s#v1.0.1",
+		"companion-module-roland-xs84h": "github:bitfocus/companion-module-roland-xs84h#v1.0.0",
+		"companion-module-rossvideo-caprica": "github:bitfocus/companion-module-rossvideo-caprica#v1.0.0",
+		"companion-module-rossvideo-nkrouter": "github:bitfocus/companion-module-rossvideo-nkrouter#v1.0.2",
+		"companion-module-rossvideo-rosstalk": "github:bitfocus/companion-module-rossvideo-rosstalk#v1.2.7",
+		"companion-module-rossvideo-xpression": "github:bitfocus/companion-module-rossvideo-xpression#v1.0.2",
+		"companion-module-sain-smart-relay": "github:bitfocus/companion-module-sain-smart-relay#v1.0.0",
+		"companion-module-sharp-tv": "github:bitfocus/companion-module-sharp-tv#v1.0.0",
+		"companion-module-showcuesystems-scs": "github:bitfocus/companion-module-showcuesystems-scs#v1.0.1",
+		"companion-module-shure-dis-ccu": "github:bitfocus/companion-module-shure-dis-ccu#v1.0.1",
+		"companion-module-shure-psm1000": "github:bitfocus/companion-module-shure-psm1000#v1.0.0",
+		"companion-module-shure-scm820": "github:bitfocus/companion-module-shure-scm820#v1.0.0",
+		"companion-module-shure-wireless": "github:bitfocus/companion-module-shure-wireless#v1.1.3",
+		"companion-module-sienna-ndimonitor": "github:bitfocus/companion-module-sienna-ndimonitor#v1.0.0",
+		"companion-module-singularlive-studio": "github:bitfocus/companion-module-singularlive-studio#v1.0.4",
+		"companion-module-slack-webhooks": "github:bitfocus/companion-module-slack-webhooks#v1.0.1",
+		"companion-module-softron-ontheairvideo": "github:bitfocus/companion-module-softron-ontheairvideo#v1.0.0",
+		"companion-module-sononum-horae": "github:bitfocus/companion-module-sononum-horae#v1.0.1",
+		"companion-module-sonoran-coyote": "github:bitfocus/companion-module-sonoran-coyote#v1.0.0",
+		"companion-module-sonos-speakers": "github:bitfocus/companion-module-sonos-speakers#v0.1.1",
+		"companion-module-sony-bravia": "github:bitfocus/companion-module-sony-bravia#v1.0.0",
+		"companion-module-sony-visca": "github:bitfocus/companion-module-sony-visca#v1.2.8",
+		"companion-module-soundcraft-ui": "github:bitfocus/companion-module-soundcraft-ui#v2.0.0",
+		"companion-module-sounddevices-pixnet": "github:bitfocus/companion-module-sounddevices-pixnet#v1.0.0",
+		"companion-module-spotify-remote": "github:bitfocus/companion-module-spotify-remote#v1.0.2",
+		"companion-module-studiocoast-vmix": "github:bitfocus/companion-module-studiocoast-vmix#v1.2.18",
+		"companion-module-tascam-bdmp1": "github:bitfocus/companion-module-tascam-bdmp1#v0.1.5",
+		"companion-module-tascam-cd": "github:bitfocus/companion-module-tascam-cd#v1.0.2",
+		"companion-module-techministry-midirelay": "github:bitfocus/companion-module-techministry-midirelay#v2.0.4",
+		"companion-module-techministry-protally": "github:bitfocus/companion-module-techministry-protally#v1.0.0",
+		"companion-module-techministry-tallyarbiter": "github:bitfocus/companion-module-techministry-tallyarbiter#v1.0.2",
+		"companion-module-teracom-tcw181b": "github:bitfocus/companion-module-teracom-tcw181b#v1.0.0",
+		"companion-module-teradek-vidiu": "github:bitfocus/companion-module-teradek-vidiu#v1.0.2",
+		"companion-module-tesla-smart": "github:bitfocus/companion-module-tesla-smart#v1.0.3",
+		"companion-module-thelightingcontroller": "github:bitfocus/companion-module-thelightingcontroller#v1.1.2",
+		"companion-module-thingm-blink1": "github:bitfocus/companion-module-thingm-blink1#v1.2.1",
+		"companion-module-tplink-kasasmartplug": "github:bitfocus/companion-module-tplink-kasasmartplug#v1.0.1",
+		"companion-module-tslproducts-umd": "github:bitfocus/companion-module-tslproducts-umd#v1.2.1",
+		"companion-module-tvone-corio": "github:bitfocus/companion-module-tvone-corio#v1.0.0",
+		"companion-module-twitch-api": "github:bitfocus/companion-module-twitch-api#v1.0.0",
+		"companion-module-ubiquiti-unifi": "github:bitfocus/companion-module-ubiquiti-unifi#v1.0.1",
+		"companion-module-vaddio-ptz": "github:bitfocus/companion-module-vaddio-ptz#v1.0.1",
+		"companion-module-vicreo-hotkey": "github:bitfocus/companion-module-vicreo-hotkey#v1.4.3",
+		"companion-module-vicreo-variablelistener": "github:bitfocus/companion-module-vicreo-variablelistener#v1.0.5",
+		"companion-module-videolan-vlc": "github:bitfocus/companion-module-videolan-vlc#v1.1.9",
+		"companion-module-visualproductions-bstation2": "github:bitfocus/companion-module-visualproductions-bstation2#v1.0.6",
+		"companion-module-vivitek-projector": "github:bitfocus/companion-module-vivitek-projector#v1.0.3",
+		"companion-module-vizio-smartcast": "github:bitfocus/companion-module-vizio-smartcast#v1.1.4",
+		"companion-module-vyv-photon": "github:bitfocus/companion-module-vyv-photon#v1.0.4",
+		"companion-module-wyrestorm-sw0402mv": "github:bitfocus/companion-module-wyrestorm-sw0402mv#v1.0.0",
+		"companion-module-xiamen-sprolink-vd-series": "github:bitfocus/companion-module-xiamen-sprolink-vd-series#v1.0.1",
+		"companion-module-yamaha-scp": "github:bitfocus/companion-module-yamaha-scp#v1.4.1",
+		"companion-module-youtube-live": "github:bitfocus/companion-module-youtube-live#v1.1.2",
+		"companion-module-zenvideo-ndirouter": "github:bitfocus/companion-module-zenvideo-ndirouter#v1.0.1",
+		"debug": "^4.2.0",
+		"elgato-stream-deck": "^4.0.1",
+		"emberplus-connection": "^0.0.4",
 		"express": "^4.16.3",
-		"find-process": "^1.4.3",
-		"flag-icon-css": "3.4.6",
-		"font-awesome": "4.7.0",
-		"fs-extra": "^7.0.0",
+		"find-process": "1.4.3",
+		"fs-extra": "^9.0.1",
 		"infinitton-idisplay": "^1.0.5",
-		"jquery": ">=3.4.0",
-		"json-parser": "^1.1.5",
-		"lodash": "^4.17.10",
-		"marked": "0.7.0",
-		"mkdirp": "^0.5.1",
-		"moment": "^2.22.2",
+		"lodash": "^4.17.20",
+		"mkdirp": "^0.5.5",
+		"moment": "^2.29.1",
 		"network": "^0.4.1",
-		"node-emberplus": "^2.5.8",
+		"node-fetch": "^2.6.1",
 		"node-rest-client": "^3.1.0",
-		"opencollective": "^1.0.3",
-		"opencollective-postinstall": "^2.0.1",
-		"osc": "^2.2.2",
-		"pace-progress": "1.0.2",
-		"perfect-scrollbar": "^1.3.0",
+		"osc": "^2.4.1",
 		"ping": "^0.2.2",
 		"pngjs": "^3.3.3",
-		"popper.js": "1.16.1",
-		"read-package-json": "^2.0.13",
-		"select2": "4.0.13",
-		"sharp": "^0.23.3",
-		"shortid": "^2.2.8",
-		"simple-line-icons": "2.4.1",
-		"socket.io": "^2.1.0",
-		"socket.io-client": "^2.1.0",
+		"sharp": "^0.27.2",
+		"shortid": "^2.2.16",
+		"socket.io": "^4.0.0",
 		"strip-ansi": "^5.2.0",
-		"tributejs": "github:peschuster/tribute",
-		"websocket": "^1.0.28"
+		"websocket": "^1.0.32"
 	},
 	"collective": {
 		"type": "donorbox",
 		"url": "https://donorbox.org/bitfocus-opensource"
 	},
 	"optionalDependencies": {
-		"@julusian/jpeg-turbo": "^0.5.4"
+		"@julusian/jpeg-turbo": "^1.0.1"
+	},
+	"resolutions": {
+		"infinitton-idisplay/node-hid": "^2.1.1",
+		"electron-builder/**/app-builder-bin": "3.5.13",
+		"**/osc/serialport": "^9.0.7"
 	}
 }
