@@ -14,8 +14,8 @@
  * disclosing the source code of your own applications.
  *
  */
-var debug  = require('debug')('lib/satellite_protocol');
-var Parser = require("binopsy");
+var debug = require('debug')('lib/satellite_protocol')
+var Parser = require('binopsy')
 
 var crc8array = [
 	0x00, 0x07, 0x0E, 0x09, 0x1C, 0x1B,
@@ -61,125 +61,101 @@ var crc8array = [
 	0xDE, 0xD9, 0xD0, 0xD7, 0xC2, 0xC5,
 	0xCC, 0xCB, 0xE6, 0xE1, 0xE8, 0xEF,
 	0xFA, 0xFD, 0xF4, 0xF3
-];
+]
 
-function protocol() {
+function protocol() {}
 
-}
+protocol.SUPPORTED_MAJOR = 1
+protocol.SUPPORTED_MINIOR = 1
 
-protocol.SUPPORTED_MAJOR = 1;
-protocol.SUPPORTED_MINIOR = 1;
+protocol.SA_HEADER_PARSER = new Parser().endianess('big').uint16('signature').uint8('command').uint16('length')
 
-protocol.SA_HEADER_PARSER = new Parser()
-	.endianess("big")
-	.uint16("signature")
-	.uint8("command")
-	.uint16("length");
-
-protocol.SCMD_ADDDEVICE_PARSER = new Parser()
-  .endianess("big")
-  .string("serialNumber", { length: 20 })
-  .uint16("deviceId");
+protocol.SCMD_ADDDEVICE_PARSER = new Parser().endianess('big').string('serialNumber', { length: 20 }).uint16('deviceId')
 
 protocol.SCMD_ADDDEVICE2_PARSER = new Parser()
-.endianess("big")
-.string("serialNumber", { length: 20 })
-.uint16("deviceId")
-.uint8('keysTotal')
-.uint8('keysPerRow')
+	.endianess('big')
+	.string('serialNumber', { length: 20 })
+	.uint16('deviceId')
+	.uint8('keysTotal')
+	.uint8('keysPerRow')
 
-protocol.SCMD_REMOVEDEVICE_PARSER = new Parser()
-  .endianess("big")
-  .uint16("deviceId");
+protocol.SCMD_REMOVEDEVICE_PARSER = new Parser().endianess('big').uint16('deviceId')
 
-protocol.SCMD_VERSION_PARSER = new Parser()
-  .endianess("big")
-	.uint8("versionMajor")
-	.uint8("versionMinior");
+protocol.SCMD_VERSION_PARSER = new Parser().endianess('big').uint8('versionMajor').uint8('versionMinior')
 
 protocol.SCMD_DRAW7272_PARSER = new Parser()
-  .endianess("big")
-	.uint16("deviceId")
-	.uint8("keyIndex")
-  .buffer("image", { length: 15552 });
+	.endianess('big')
+	.uint16('deviceId')
+	.uint8('keyIndex')
+	.buffer('image', { length: 15552 })
 
-protocol.SCMD_BUTTON_PARSER = new Parser()
-  .endianess("big")
-	.uint16("deviceId")
-	.uint8("keyIndex")
-  .uint8("state");
+protocol.SCMD_BUTTON_PARSER = new Parser().endianess('big').uint16('deviceId').uint8('keyIndex').uint8('state')
 
-protocol.SCMD_BRIGHTNESS_PARSER = new Parser()
-  .endianess("big")
-	.uint16("deviceId")
-  .uint8("percent");
+protocol.SCMD_BRIGHTNESS_PARSER = new Parser().endianess('big').uint16('deviceId').uint8('percent')
 
-
-
-protocol.SCMD_PING = 0x01;
-protocol.SCMD_PONG = 0x02;
-protocol.SCMD_VERSION = 0x03;
-protocol.SCMD_ADDDEVICE = 0x10;
-protocol.SCMD_REMOVEDEVICE = 0x11;
-protocol.SCMD_ADDDEVICE2 = 0x12;
-protocol.SCMD_BUTTON = 0x20;
-protocol.SCMD_DRAW7272 = 0x30;
-protocol.SCMD_BRIGHTNESS = 0x31;
-
-
+protocol.SCMD_PING = 0x01
+protocol.SCMD_PONG = 0x02
+protocol.SCMD_VERSION = 0x03
+protocol.SCMD_ADDDEVICE = 0x10
+protocol.SCMD_REMOVEDEVICE = 0x11
+protocol.SCMD_ADDDEVICE2 = 0x12
+protocol.SCMD_BUTTON = 0x20
+protocol.SCMD_DRAW7272 = 0x30
+protocol.SCMD_BRIGHTNESS = 0x31
 
 protocol.calcCRC = function (buffer) {
-	var crc = 0;
+	var crc = 0
 
 	for (var i = 0; i < buffer.length; ++i) {
-		crc = crc8array[crc ^ buffer[i]];
+		crc = crc8array[crc ^ buffer[i]]
 	}
 
-	return crc;
+	return crc
 }
 
-protocol.readHeader = function(buffer) {
-	var self = this;
+protocol.readHeader = function (buffer) {
+	var self = this
 
 	// Not enough data
-	if (buffer.length < 5) return false;
+	if (buffer.length < 5) return false
 
-	if (buffer.readUInt16BE(0) !== 0x5A7E) { // SATE(lite)
-		return -1; // out of sync
+	if (buffer.readUInt16BE(0) !== 0x5a7e) {
+		// SATE(lite)
+		return -1 // out of sync
 	}
 
-	obj = protocol.SA_HEADER_PARSER.parse(buffer);
+	obj = protocol.SA_HEADER_PARSER.parse(buffer)
 
-	return obj;
-};
+	return obj
+}
 
-protocol.sendPacket = function(socket, command, data) {
-	var self = this;
+protocol.sendPacket = function (socket, command, data) {
+	var self = this
 
 	if (data === undefined) {
-		data = new Buffer('');
+		data = Buffer.from('')
 	}
 
 	if (typeof data !== 'object' || !(data instanceof Buffer)) {
-		data = new Buffer(data);
+		data = Buffer.from(data)
 	}
 
-	var buffer = new Buffer(5 + data.length + 1);
+	var buffer = Buffer.alloc(5 + data.length + 1)
 
 	var header = {
-		signature: 0x5A7E,
+		signature: 0x5a7e,
 		command: command,
 		length: data.length
-	};
+	}
 
-	protocol.SA_HEADER_PARSER.serialize(header, buffer);
-	data.copy(buffer, 5);
+	protocol.SA_HEADER_PARSER.serialize(header, buffer)
+	data.copy(buffer, 5)
 
-	var crc = self.calcCRC(data);
+	var crc = self.calcCRC(data)
 
-	buffer[5+data.length] = crc;
+	buffer[5 + data.length] = crc
 
-	socket.write(buffer);
-};
+	socket.write(buffer)
+}
 
-module.exports = protocol;
+module.exports = protocol
