@@ -15,9 +15,9 @@
  *
  */
 
-
-var debug = require('debug')('lib/db');
-var fs    = require('fs-extra');
+var debug = require('debug')('lib/db')
+var fs = require('fs-extra')
+var path = require('path')
 
 /**
 	Simpel KVS som forventer at all data har plass i ram.
@@ -33,157 +33,163 @@ var fs    = require('fs-extra');
 		* db_save - Lagrer db fra minne til fil. Svarer med db_saved. (se over)
 */
 
-var lastsave = 0;
-var saveInterval = 4000; // Minimum 4 seconds between each save
-var dirty = false;
+var lastsave = 0
+var saveInterval = 4000 // Minimum 4 seconds between each save
+var dirty = false
+var saving = false
 
-module.exports = exports = function (system,cfgDir) {
-	return new db(system, cfgDir);
-};
+module.exports = exports = function (system, cfgDir) {
+	return new db(system, cfgDir)
+}
 
-function db(system,cfgDir) {
-	debug('new(db)');
-	var self = this;
-	self.system = system;
-	self.db = {};
+function db(system, cfgDir) {
+	debug('new(db)')
+	var self = this
+	self.system = system
+	self.db = {}
+
+	var pathDb = path.join(cfgDir, 'db')
+	var pathDbTmp = path.join(cfgDir, 'db.tmp')
+	var pathDbBak = path.join(cfgDir, 'db.bak')
 
 	try {
-		var data = fs.readFileSync(cfgDir + '/db');
+		var data = fs.readFileSync(pathDb)
 
-		self.db = JSON.parse(data);
-		debug('db loaded');
+		self.db = JSON.parse(data)
+		debug('db loaded')
 
-		var changed_after_load = false;
+		var changed_after_load = false
 		// db defaults
 		if (self.db.userconfig === undefined) {
-			self.db.userconfig = {};
-			changed_after_load = true;
+			self.db.userconfig = {}
+			changed_after_load = true
 		}
 
 		// is page up 1->2 or 2->1?
 		if (self.db.userconfig.page_direction_flipped === undefined) {
-			self.db.userconfig.page_direction_flipped = false;
-			changed_after_load = true;
+			self.db.userconfig.page_direction_flipped = false
+			changed_after_load = true
 		}
 
-		system.emit('db_loaded', self.db);
+		system.emit('db_loaded', self.db)
 
 		if (changed_after_load === true) {
-			debug('config changed by default values after load, saving.');
-			system.emit('db_save');
+			debug('config changed by default values after load, saving.')
+			system.emit('db_save')
 		}
-
 	} catch (err) {
-
 		if (err.code == 'ENOENT') {
-			debug("readFile(db)","Couldnt read db, loading {}");
-			system.emit('db_loaded', {});
+			debug('readFile(db)', 'Couldnt read db, loading {}')
+			system.emit('db_loaded', {})
 		} else {
-			throw err;
+			throw err
 		}
-
 	}
 
-	system.on("db_all", function(cb) {
-		debug("db_all(): returning all database values");
+	system.on('db_all', function (cb) {
+		debug('db_all(): returning all database values')
 		if (typeof cb == 'function') {
-			cb(self.db);
+			cb(self.db)
 		}
-	});
+	})
 
 	system.on('db_set', function (key, value) {
-		debug('db_set(' + key + ', ' + value + ')');
-		self.db[key] = value;
-	});
+		debug('db_set(' + key + ', ' + value + ')')
+		self.db[key] = value
+	})
 
 	system.on('db_del', function (key) {
-		debug('db_del(' + key + ')');
-		delete self.db[key];
-	});
+		debug('db_del(' + key + ')')
+		delete self.db[key]
+	})
 
 	system.on('db_set_multiple', function (keyvalueobj) {
-		debug('db_set_multiple:');
+		debug('db_set_multiple:')
 		for (var key in keyvalueobj) {
-			debug('db_set(' + key + ',' + keyvalueobj[key] + ')');
-			self.db[key] = keyvalueobj[key];
+			debug('db_set(' + key + ',' + keyvalueobj[key] + ')')
+			self.db[key] = keyvalueobj[key]
 		}
-	});
-
-
-	/*system.on('db_get', function (key, cb) {
-		debug('db_get(' + key + ')');
-		setImmediate(function () {
-			debug('db_get[RETURN](' + key + '): '+self.db[key]);
-			system.emit(cb, self.db[key]);
-		});
-	});*/
+	})
 
 	system.on('db_get', function (key, cb) {
-		debug('db_get(' + key + ')');
-		cb(self.db[key]);
-	});
+		debug('db_get(' + key + ')')
+		cb(self.db[key])
+	})
 	system.on('db_get_multiple', function (keys, cb) {
 		if (typeof keys != 'object' || typeof keys.length == 'undefined') {
-			throw new Error('keys is not an array');
+			throw new Error('keys is not an array')
 		}
-		cb(keys.map(function (key) {
-			return self.db[key];
-		}));
-	});
+		cb(
+			keys.map(function (key) {
+				return self.db[key]
+			})
+		)
+	})
+
+	async function doSave() {
+		const jsonStr = JSON.stringify(self.db)
+
+		try {
+			await fs.copy(pathDb, pathDbBak)
+		} catch (err) {
+			debug('db_save', 'Error making backup of config: ' + err)
+		}
+
+		try {
+			await fs.writeFile(pathDbTmp, jsonStr)
+		} catch (err) {
+			debug('db_save', 'Error saving: ' + err)
+			throw 'Error saving: ' + err
+		}
+
+		debug('db_save', 'written')
+
+		try {
+			await fs.rename(pathDbTmp, pathDb)
+		} catch (err) {
+			debug('db_save', 'Error renaming db.tmp: ' + err)
+			throw 'Error renaming db.tmp: ' + err
+		}
+
+		debug('db_save', 'renamed')
+		system.emit('db_saved', null)
+	}
 
 	system.on('db_save', function () {
+		if (!saving && Date.now() - lastsave > saveInterval) {
+			debug('db_save', 'begin')
 
-		if (Date.now() - lastsave > saveInterval) {
-			debug("db_save","begin");
+			dirty = false
+			lastsave = Date.now()
+			saving = true
 
-			dirty = false;
-
-			lastsave = Date.now();
-
-			fs.copy(cfgDir + '/db', cfgDir + '/db.bak', function (err) {
-				if (err) {
-					debug('db_save', 'Error making backup of config: ' + err);
-				}
-
-				fs.writeFile(cfgDir + '/db.tmp', JSON.stringify(self.db), function (err) {
-					if (err) {
-						debug('db_save', 'Error saving: ' + err);
-						system.emit('db_saved', err);
-						return;
+			doSave()
+				.catch((err) => {
+					try {
+						self.system.emit('log', 'CORE(cb)', 'error', err)
+					} catch (err2) {
+						// make sure to not throw an error here
+						debug('db_save', 'Error reporting save failure: ' + err2)
 					}
-
-					debug("db_save","written");
-
-					fs.rename(cfgDir + '/db.tmp', cfgDir + '/db', function(err) {
-
-						if (err) {
-							self.system.emit('log', 'CORE(cb)', 'error', 'db.tmp->db failed: ' + err);
-						}
-						else {
-							debug('db_save','renamed');
-							system.emit('db_saved', null);
-						}
-
-					});
-
-				});
-
-			});
-
+				})
+				.then(() => {
+					// This will run even if the catch caught an error
+					saving = false
+				})
 		} else {
-			dirty = true;
+			dirty = true
 		}
-	});
+	})
 
 	// Do a save sometime within the next 10 seconds
 	system.on('db_dirty', function () {
-		dirty = true;
-	});
+		dirty = true
+	})
 
 	// If last db_save was not handeled because of throttling, do it now
 	setInterval(function () {
 		if (Date.now() - lastsave > saveInterval && dirty) {
-			system.emit('db_save');
+			system.emit('db_save')
 		}
-	}, 4000);
-};
+	}, 4000)
+}
