@@ -15,89 +15,63 @@
  *
  */
 
-var debug  = require('debug')('lib/help');
-var path   = require('path');
-var fs     = require('fs');
-var marked = require('marked');
+var debug = require('debug')('lib/help')
+var fs = require('fs')
 
 function help(system) {
-	var self = this;
-
-	/**
-	 * Reads the Markdown file located in /documentation and returns it as HTML.
-	 * 
-	 * @param markdownFilename
-	 */
-	function getDocumentationFile(markdownFilename) {
-
-		// Prevent directory traversal
-		markdownFilename = markdownFilename.replace(/(\.\.|\/)/g, '');
-		var path = require('app-root-path') + '/documentation/' + markdownFilename;
-
-		if (fs.existsSync(path)) {
-			try {
-				var help = fs.readFileSync(path);
-				help = marked(help.toString(), {
-					baseUrl      : '/int/documentation/',
-					headerPrefix : 'header-',
-				});
-				return help;
-			} catch (e) {
-				debug('Error loading help ' + path);
-				debug(e);
-			}
-		}
-
-		return null;
-	
-	}
-
+	var self = this
 
 	system.on('http_req', function (req, res, done) {
 		// Watch for HTTP requests made to /int/ (which emits 'http_req').
-		var match;
+		var match
 
 		// Return any asset made to /int/documentation/
-		if (match = req.url.match(/^\/documentation\/(.+?)(\?.+)?$/)) {
-			var path = require('app-root-path') + '/documentation';
-			var file = match[1].replace(/\.\.+/g, '');
+		if ((match = req.url.match(/^\/documentation\/(.+?)(\?.+)?$/))) {
+			var path = require('app-root-path') + '/documentation'
+			var file = match[1].replace(/\.\.+/g, '')
 
-			if (file.match(/\.(jpe?g|gif|png|pdf)$/) && fs.existsSync(path +'/'+ file)) {
-				done();
-				res.sendFile(path +'/'+ file);
+			if (file.match(/\.(jpe?g|gif|png|pdf)$/) && fs.existsSync(path + '/' + file)) {
+				done()
+				res.sendFile(path + '/' + file)
 			}
 		}
-	});
-
+	})
 
 	system.on('io_connect', function (client) {
-
-		client.on('get_help', function (req) {
+		client.on('get_help_md', function (req, answer) {
 			// Return the rendered HTML for the requested Markdown file.
 			// req will look like { file:'file.md' }.
 
 			// Pass the filename requested back to the response so the receiver can watch for
 			//  their response in case two different requests were made at the same time.
 			var resp = {
-				error : true,
-				file  : req.file,
-				html  : '',
-			};
-
-			if (req.file !== undefined) {
-				var html;
-				if ((html = getDocumentationFile(req.file)) !== null) {
-					resp.error = false;
-					resp.html  = html;
-				}
-			} else {
-				debug('Invalid get_help request');
+				error: true,
+				file: req.file,
+				markdown: '',
 			}
 
-			client.emit('get_help:result', resp);
-		});
-	});
+			if (req.file !== undefined) {
+				// Prevent directory traversal
+				const markdownFilename = req.file.replace(/(\.\.|\/)/g, '')
+				const path = require('app-root-path') + '/documentation/' + markdownFilename
 
+				if (fs.existsSync(path)) {
+					try {
+						resp.markdown = fs.readFileSync(path).toString()
+						resp.baseUrl = '/int/documentation/'
+						resp.error = false
+					} catch (e) {
+						debug('Error loading help ' + path)
+						debug(e)
+					}
+				}
+			} else {
+				debug('Invalid get_help request')
+			}
+
+			answer(resp)
+		})
+	})
 }
 
-module.exports = help;
+module.exports = help
