@@ -15,45 +15,34 @@
  *
  */
 
-var util    = require("util");
-var _io     = require('socket.io');
-var debug   = require('debug')('lib/io');
-var system;
+var _io = require('socket.io')
+var debug = require('debug')('lib/io')
 
-function io(_system, http) {
-	var self = this;
-	system = _system;
-	self.modules = {};
-	self.system = _system;
+exports = module.exports = function (system, http) {
+	const io = _io(http, {
+		allowEIO3: true,
+		cors: {
+			// Allow everything
+			origin: (o, cb) => cb(null, o),
+			credentials: true,
+		},
+	})
 
-	_io.call(self, http);
-
-	self.system.on('io_get', function (cb) {
+	system.on('io_get', function (cb) {
 		if (typeof cb == 'function') {
-			cb(self);
+			cb(io)
 		}
-	});
+	})
 
-	self.initIO();
+	io.on('connect', function (client) {
+		debug('io-connect')
+
+		system.emit('skeleton-info-info', function (info) {
+			client.emit('skeleton-info', info)
+		})
+
+		system.emit('io_connect', client)
+	})
+
+	return io
 }
-util.inherits(io, _io);
-
-io.prototype.initIO = function() {
-	var self = this;
-
-	self.on('connect', function (client) {
-
-		debug('io-connect');
-
-		system.emit('skeleton-info-info', function(hash) {
-			client.emit('skeleton-info', hash);
-		});
-
-		system.emit('io_connect', client);
-
-	});
-};
-
-exports = module.exports = function (_system, http) {
-	return new io(_system, http);
-};
