@@ -15,42 +15,42 @@
  *
  */
 
-var debug   = require('debug')('lib/log');
+var debug = require('debug')('lib/log')
 
-function log(system,io) {
-	var self = this;
-	self.system = system;
-	self.io = io;
-	self.history = [
-		[Date.now(), 'log', 'info', 'Application started']
-	];
+function log(system, io) {
+	var self = this
+	self.system = system
+	self.io = io
+	self.history = [[Date.now(), 'log', 'info', 'Application started']]
 
-	self.system.on('log', function(source, level, message) {
-		var now = Date.now();
-		io.emit('log',now,source,level,message);
-		self.history.push([now,source,level,message])
-		if (self.history.length > 500) {
-			self.history.shift();
-		}
-	});
-
-	system.on('io_connect', function(client) {
-		client.on('log_clear', function() {
-			client.broadcast.emit('log_clear');
-			self.history = [];
-			self.io.emit('log', Date.now(), 'log', 'info', 'Log cleared');
-		});
-		client.on('log_catchup', function() {
-			for (var n in self.history) {
-				var arr = self.history[n];
-				client.emit('log', arr[0], arr[1], arr[2], arr[3]);
+	self.system.on('log', function (source, level, message) {
+		if (level) {
+			var now = Date.now()
+			io.emit('log', now, source, level, message)
+			self.history.push([now, source, level, message])
+			if (self.history.length > 500) {
+				self.history.shift()
 			}
-		});
-	});
+		}
+	})
 
-	return self;
+	system.on('io_connect', function (client) {
+		client.on('log_clear', function () {
+			client.broadcast.emit('log_clear')
+			self.history = []
+			self.io.emit('log', Date.now(), 'log', 'info', 'Log cleared')
+		})
+		client.on('log_catchup', function () {
+			for (var n in self.history) {
+				var arr = self.history[n]
+				client.emit('log', arr[0], arr[1], arr[2], arr[3])
+			}
+		})
+	})
+
+	return self
 }
 
-exports = module.exports = function (system,io) {
-	return new log(system,io);
-};
+exports = module.exports = function (system, io) {
+	return new log(system, io)
+}
