@@ -15,103 +15,101 @@
  *
  */
 
-var debug   = require('debug')('lib/osc');
-var OSC     = require('osc')
+var debug = require('debug')('lib/osc')
+var OSC = require('osc')
 
 function osc(system) {
-	var self = this;
-	self.ready = true;
-	self.system = system;
+	var self = this
+	self.ready = true
+	self.system = system
 
 	self.listener = new OSC.UDPPort({
-		localAddress: "0.0.0.0",
+		localAddress: '0.0.0.0',
 		localPort: 12321,
 		broadcast: true,
-		metadata: true
-	});
+		metadata: true,
+	})
 
-	self.listener.open();
+	self.listener.open()
 
-	self.listener.on("ready", function () {
-		self.ready = true;
-	});
+	self.listener.on('ready', function () {
+		self.ready = true
+	})
 
-	self.listener.on("message", function (message) {
-		var a = message.address.split("/");
-		if (message.address.match(/^\/press\/bank\/\d+\/\d+$/)) {
+	self.listener.on('message', function (message) {
+		try {
+			var a = message.address.split('/')
+			if (message.address.match(/^\/press\/bank\/\d+\/\d+$/)) {
+				if (message.args.length == 0) {
+					debug('Got /press/bank/ (trigger)', parseInt(a[3]), 'button', parseInt(a[4]))
+					system.emit('bank_pressed', parseInt(a[3]), parseInt(a[4]), true)
 
-			if (message.args.length == 0) {
-
-				debug("Got /press/bank/ (trigger)",parseInt(a[3]),"button",parseInt(a[4]));
-				system.emit('bank_pressed', parseInt(a[3]), parseInt(a[4]), true);
-
-				setTimeout(function (){
-					debug("Auto releasing /press/bank/ (trigger)",parseInt(a[3]),"button",parseInt(a[4]));
-					system.emit('bank_pressed', parseInt(a[3]), parseInt(a[4]), false);
-				}, 20);
-
-			} else {
-
-				if (message.args[0].type == 'i' && message.args[0].value == '1') {
-					debug("Got /press/bank/ (press) for bank", parseInt(a[3]), "button", parseInt(a[4]));
-					system.emit('bank_pressed', parseInt(a[3]), parseInt(a[4]), true);
+					setTimeout(function () {
+						debug('Auto releasing /press/bank/ (trigger)', parseInt(a[3]), 'button', parseInt(a[4]))
+						system.emit('bank_pressed', parseInt(a[3]), parseInt(a[4]), false)
+					}, 20)
+				} else {
+					if (message.args[0].type == 'i' && message.args[0].value == '1') {
+						debug('Got /press/bank/ (press) for bank', parseInt(a[3]), 'button', parseInt(a[4]))
+						system.emit('bank_pressed', parseInt(a[3]), parseInt(a[4]), true)
+					} else if (message.args[0].type == 'i' && message.args[0].value == '0') {
+						debug('Got /press/bank/ (release) for bank', parseInt(a[3]), 'button', parseInt(a[4]))
+						system.emit('bank_pressed', parseInt(a[3]), parseInt(a[4]), false)
+					}
 				}
-				else if (message.args[0].type == 'i' && message.args[0].value == '0') {
-					debug("Got /press/bank/ (release) for bank", parseInt(a[3]), "button", parseInt(a[4]));
-					system.emit('bank_pressed', parseInt(a[3]), parseInt(a[4]), false);
+			} else if (message.address.match(/^\/style\/bgcolor\/\d+\/\d+$/)) {
+				if (message.args.length > 2) {
+					var r = message.args[0].value
+					var g = message.args[1].value
+					var b = message.args[2].value
+					if (typeof r === 'number' && typeof g === 'number' && typeof b === 'number') {
+						var col = rgb(r, g, b)
+						debug('Got /style/bgcolor', parseInt(a[3]), 'button', parseInt(a[4]))
+						system.emit('bank_set_key', parseInt(a[3]), parseInt(a[4]), 'bgcolor', col)
+						system.emit('graphics_bank_invalidate', parseInt(a[3]), parseInt(a[4]))
+					}
 				}
-
-			}
-		} else if (message.address.match(/^\/style\/bgcolor\/\d+\/\d+$/)) {
-			if (message.args.length > 2) {
-				var r = message.args[0].value;
-				var g = message.args[1].value;
-				var b = message.args[2].value;
-				if (typeof r === "number" && typeof g === "number" && typeof b === "number") {
-					var col = rgb(r, g, b);
-					debug("Got /style/bgcolor", parseInt(a[3]), "button", parseInt(a[4]))
-					system.emit('bank_set_key', parseInt(a[3]), parseInt(a[4]), 'bgcolor', col);
-					system.emit('graphics_bank_invalidate', parseInt(a[3]), parseInt(a[4]));
+			} else if (message.address.match(/^\/style\/color\/\d+\/\d+$/)) {
+				if (message.args.length > 2) {
+					var r = message.args[0].value
+					var g = message.args[1].value
+					var b = message.args[2].value
+					if (typeof r === 'number' && typeof g === 'number' && typeof b === 'number') {
+						var col = rgb(r, g, b)
+						debug('Got /style/color', parseInt(a[3]), 'button', parseInt(a[4]))
+						system.emit('bank_set_key', parseInt(a[3]), parseInt(a[4]), 'color', col)
+						system.emit('graphics_bank_invalidate', parseInt(a[3]), parseInt(a[4]))
+					}
 				}
-			}
-		} else if (message.address.match(/^\/style\/color\/\d+\/\d+$/)) {
-			if (message.args.length > 2) {
-				var r = message.args[0].value;
-				var g = message.args[1].value;
-				var b = message.args[2].value;
-				if (typeof r === "number" && typeof g === "number" && typeof b === "number") {
-					var col = rgb(r, g, b);
-					debug("Got /style/color", parseInt(a[3]), "button", parseInt(a[4]))
-					system.emit('bank_set_key', parseInt(a[3]), parseInt(a[4]), 'color', col);
-					system.emit('graphics_bank_invalidate', parseInt(a[3]), parseInt(a[4]));
-				}
-			}
-		} else if (message.address.match(/^\/style\/text\/\d+\/\d+$/)) {
-			if (message.args.length > 0) {
-				var text = message.args[0].value;
-				if (typeof text === "string") {
-					debug("Got /style/text", parseInt(a[3]), "button", parseInt(a[4]))
-					system.emit('bank_set_key', parseInt(a[3]), parseInt(a[4]), 'text', text);
-					system.emit('graphics_bank_invalidate', parseInt(a[3]), parseInt(a[4]));
+			} else if (message.address.match(/^\/style\/text\/\d+\/\d+$/)) {
+				if (message.args.length > 0) {
+					var text = message.args[0].value
+					if (typeof text === 'string') {
+						debug('Got /style/text', parseInt(a[3]), 'button', parseInt(a[4]))
+						system.emit('bank_set_key', parseInt(a[3]), parseInt(a[4]), 'text', text)
+						system.emit('graphics_bank_invalidate', parseInt(a[3]), parseInt(a[4]))
+					}
 				}
 			}
+		} catch (error) {
+			system.emit('log', 'osc', 'warn', 'OSC Error: ' + error)
 		}
-	});
+	})
 
-	function rgb(r,g,b) {
-		return (
-			((r & 0xff) << 16) |
-			((g & 0xff) << 8) |
-			(b & 0xff)
-		);
-	};
+	function rgb(r, g, b) {
+		return ((r & 0xff) << 16) | ((g & 0xff) << 8) | (b & 0xff)
+	}
 
-	self.system.on('osc_send', function(host, port, path, args) {
-		self.listener.send({
-			address: path,
-			args: args
-		}, host, port);
-	});
+	self.system.on('osc_send', function (host, port, path, args) {
+		self.listener.send(
+			{
+				address: path,
+				args: args,
+			},
+			host,
+			port
+		)
+	})
 
 	/*
 		Example usage of bundle scheduled after 60 seconds:
@@ -131,17 +129,20 @@ function osc(system) {
 			}
 		]);
 	*/
-	self.system.on('osc_send_bundle', function(host, port, time, bundle) {
-		self.listener.send({
-			timeTag: OSC.timeTag(time),
-			packets: bundle
-		}, host, port);
-	});
+	self.system.on('osc_send_bundle', function (host, port, time, bundle) {
+		self.listener.send(
+			{
+				timeTag: OSC.timeTag(time),
+				packets: bundle,
+			},
+			host,
+			port
+		)
+	})
 
-
-	return self;
+	return self
 }
 
 exports = module.exports = function (system) {
-	return new osc(system);
-};
+	return new osc(system)
+}
