@@ -15,83 +15,79 @@
  *
  */
 
-var debug   = require('debug')('lib/page');
-var system;
+var debug = require('debug')('lib/page')
+var system
 
 function page(system) {
-	var self = this;
+	var self = this
 
-	self.page = {};
+	self.page = {}
 
-	system.emit('db_get', 'page', function(config) {
-		self.page = config;
+	system.emit('db_get', 'page', function (config) {
+		self.page = config
 
 		// Default values
 		if (self.page === undefined) {
-			self.page = {};
+			self.page = {}
 			for (var n = 1; n <= 99; n++) {
-				if (self.page[''+n] === undefined) {
-					self.page[''+n] = {
-						name: 'PAGE'
-					};
+				if (self.page['' + n] === undefined) {
+					self.page['' + n] = {
+						name: 'PAGE',
+					}
 				}
 			}
 		}
-
-	});
+	})
 
 	system.on('page_set_noredraw', function (page, name) {
-		debug('NR: Set page ' + page + ' to ', name);
-		self.page[page] = name;
-		self.io.emit('set_page', page, name);
-	});
+		debug('NR: Set page ' + page + ' to ', name)
+		self.page[page] = name
+		self.io.emit('set_page', page, name)
+	})
 
-	system.on('page_set', function (page, value, io) {
+	system.on('page_set', function (page, value) {
+		debug('Set page ' + page + ' to ', value)
+		self.page[page] = value
 
-		debug('Set page ' + page + ' to ', value);
-		self.page[page] = value;
+		self.io.emit('set_page', page, value)
 
-		if (io !== undefined) {
-			io.broadcast.emit('set_page', page, value);
-		} else {
-			self.io.emit('set_page', page, value);
-		}
+		system.emit('db_set', 'page', self.page)
+		system.emit('page_update', page, value)
+		system.emit('db_save')
+	})
 
-		system.emit('db_set', 'page', self.page);
-		system.emit('page_update', page, value);
-		system.emit('db_save');
-	});
-
-	system.on('get_page', function(cb) {
-		cb(self.page);
-	});
+	system.on('get_page', function (cb) {
+		cb(self.page)
+	})
 
 	system.emit('io_get', function (io) {
-		self.io = io;
+		self.io = io
 
 		system.on('io_connect', function (socket) {
-
-			debug('socket ' + socket.id + ' connected');
+			debug('socket ' + socket.id + ' connected')
 
 			socket.on('set_page', function (key, value) {
-				debug('socket: set_page ' + key, value);
-				system.emit('page_set', key, value, socket);
-			});
+				debug('socket: set_page ' + key, value)
+				system.emit('page_set', key, value)
+			})
 
-			socket.on('get_page_all', function () {
-				debug("socket: get_page_all");
-				socket.emit('get_page_all', self.page);
-			});
+			socket.on('get_page_all', function (answer) {
+				debug('socket: get_page_all')
+
+				if (typeof answer === 'function') {
+					answer(self.page)
+				} else {
+					socket.emit('get_page_all', self.page)
+				}
+			})
 
 			socket.on('disconnect', function () {
-				debug('socket ' + socket.id + ' disconnected');
-			});
-
-		});
-	});
-
+				debug('socket ' + socket.id + ' disconnected')
+			})
+		})
+	})
 }
 
 module.exports = function (system) {
-	return new page(system);
-};
+	return new page(system)
+}
