@@ -15,98 +15,103 @@
  *
  */
 
-var system;
-var debug   = require('debug')('lib/preset');
-var shortid = require('shortid');
+var system
+var debug = require('debug')('lib/preset')
+var shortid = require('shortid')
 
 function preset(system) {
-	var self = this;
+	var self = this
 
-	self.system = system;
-	self.presets = {};
+	self.system = system
+	self.presets = {}
 
-	self.system.emit('io_get', function(io) {
-		self.io = io;
-		self.system.on('io_connect', function(client) {
+	self.system.emit('io_get', function (io) {
+		self.io = io
+		self.system.on('io_connect', function (client) {
+			function sendResult(answer, name, ...args) {
+				if (typeof answer === 'function') {
+					answer(...args)
+				} else {
+					client.emit(name, ...args)
+				}
+			}
 
-			client.on('get_presets', function() {
-				client.emit('get_presets:result', self.presets);
-			});
+			client.on('get_presets', function (answer) {
+				sendResult(answer, 'get_presets:result', self.presets)
+			})
 
 			client.on('preset_drop', function (instance, config, page, bank) {
-
 				if (config.actions !== undefined) {
 					for (var i = 0; i < config.actions.length; ++i) {
-						config.actions[i].id = shortid.generate();
-						config.actions[i].instance = instance;
-						config.actions[i].label = instance + ':' + config.actions[i].action;
+						config.actions[i].id = shortid.generate()
+						config.actions[i].instance = instance
+						config.actions[i].label = instance + ':' + config.actions[i].action
 					}
 				} else {
-					config.actions = [];
+					config.actions = []
 				}
 
 				if (config.release_actions !== undefined) {
 					for (var i = 0; i < config.release_actions.length; ++i) {
-						config.release_actions[i].id = shortid.generate();
-						config.release_actions[i].instance = instance;
-						config.release_actions[i].label = instance + ':' + config.release_actions[i].action;
+						config.release_actions[i].id = shortid.generate()
+						config.release_actions[i].instance = instance
+						config.release_actions[i].label = instance + ':' + config.release_actions[i].action
 					}
 				} else {
-					config.release_actions = [];
+					config.release_actions = []
 				}
 
 				if (config.feedbacks !== undefined) {
 					for (var i = 0; i < config.feedbacks.length; ++i) {
-						config.feedbacks[i].id = shortid.generate();
-						config.feedbacks[i].instance_id = instance;
+						config.feedbacks[i].id = shortid.generate()
+						config.feedbacks[i].instance_id = instance
 					}
 				} else {
-					config.feedbacks = [];
+					config.feedbacks = []
 				}
 
-				config.config = config.bank;
-				delete config.bank;
+				config.config = config.bank
+				delete config.bank
 
 				self.system.emit('import_bank', page, bank, config, function () {
-					client.emit('preset_drop:result', null, 'ok');
-				});
-			});
-
-		});
-	});
+					client.emit('preset_drop:result', null, 'ok')
+				})
+			})
+		})
+	})
 
 	self.system.on('instance_enable', function (id, state) {
 		if (state === false) {
-			delete self.presets[id];
-			self.io.emit('presets_delete', id);
+			delete self.presets[id]
+			self.io.emit('presets_delete', id)
 		}
-	});
+	})
 
 	self.system.on('instance_delete', function (id) {
-		delete self.presets[id];
-		self.io.emit('presets_delete', id);
-	});
+		delete self.presets[id]
+		self.io.emit('presets_delete', id)
+	})
 
 	self.system.on('preset_instance_definitions_set', function (instance, presets) {
-		self.presets[instance.id] = presets;
-		self.io.emit('presets_update', instance.id, presets);
-	});
+		self.presets[instance.id] = presets
+		self.io.emit('presets_update', instance.id, presets)
+	})
 
 	system.on('variable_instance_label_rename', function (labelFrom, labelTo, id) {
 		if (self.presets[id] !== undefined) {
 			system.emit('instance_get', id, function (instance) {
 				if (instance !== undefined) {
-					instance.label = labelTo;
-					debug('Updating presets for instance ' + labelTo);
-					instance.setPresetDefinitions(self.presets[id]);
+					instance.label = labelTo
+					debug('Updating presets for instance ' + labelTo)
+					instance.setPresetDefinitions(self.presets[id])
 				}
-			});
+			})
 		}
-	});
+	})
 
-	return self;
+	return self
 }
 
 exports = module.exports = function (system) {
-	return new preset(system);
-};
+	return new preset(system)
+}
