@@ -15,93 +15,93 @@
  *
  */
 
-var debug   = require('debug')('lib/rest');
-var Client  = require('node-rest-client').Client;
+var debug = require('debug')('lib/rest')
+var Client = require('node-rest-client').Client
 
 function rest(system) {
-	var self = this;
-	self.system = system;
+	var self = this
+	self.system = system
 
-	system.on('rest_get', function(url, cb, extra_headers, extra_args) {
+	system.on('rest_get', function (url, cb, extra_headers, extra_args) {
+		debug('making request:', url)
 
-		debug('making request:', url);
-
-		var client = new Client(extra_args);
+		var client = new Client(extra_args)
 
 		var args = {
-			headers: { "Content-Type": "application/json" }
-		};
+			headers: { 'Content-Type': 'application/json' },
+		}
 
 		if (extra_headers !== undefined) {
 			for (var header in extra_headers) {
-				args.headers[header] = extra_headers[header];
+				args.headers[header] = extra_headers[header]
 			}
 		}
 
-		client.get(url, args, function (data, response) {
-			cb(null, { data: data, response: response });
-		}).on('error', function(error) {
-			debug('error response:', error);
-			cb(true, { error: error });
-		});
+		client
+			.get(url, args, function (data, response) {
+				cb(null, { data: data, response: response })
+			})
+			.on('error', function (error) {
+				debug('error response:', error)
+				cb(true, { error: error })
+			})
+	})
 
-	});
+	system.on('rest', function (url, data, cb, extra_headers, extra_args) {
+		debug('making request:', url, data)
 
-	system.on('rest', function(url, data, cb, extra_headers, extra_args) {
-
-		debug('making request:', url, data);
-
-		var client = new Client(extra_args);
-
-		var args = {
-			data: data,
-			headers: { "Content-Type": "application/json" }
-		};
-
-		if (extra_headers !== undefined) {
-			for (var header in extra_headers) {
-				args.headers[header] = extra_headers[header];
-			}
-		}
-
-		client.post(url, args, function (data, response) {
-			cb(null, { data: data, response: response });
-		}).on('error', function(error) {
-			debug('error response:', error);
-			cb(true, { error: error });
-		});
-
-	});
-
-	system.on('rest_put', function(url, data, cb, extra_headers, extra_args) {
-
-		debug('making request:', url, data);
-
-		var client = new Client(extra_args);
+		var client = new Client(extra_args)
 
 		var args = {
 			data: data,
-			headers: { "Content-Type": "application/json" }
-		};
+			headers: { 'Content-Type': 'application/json' },
+		}
 
 		if (extra_headers !== undefined) {
 			for (var header in extra_headers) {
-				args.headers[header] = extra_headers[header];
+				args.headers[header] = extra_headers[header]
 			}
 		}
 
-		client.put(url, args, function (data, response) {
-			cb(null, { data: data, response: response });
-		}).on('error', function(error) {
-			debug('error response:', error);
-			cb(true, { error: error });
-		});
+		client
+			.post(url, args, function (data, response) {
+				cb(null, { data: data, response: response })
+			})
+			.on('error', function (error) {
+				debug('error response:', error)
+				cb(true, { error: error })
+			})
+	})
 
-	});
+	system.on('rest_put', function (url, data, cb, extra_headers, extra_args) {
+		debug('making request:', url, data)
 
-	return self;
+		var client = new Client(extra_args)
+
+		var args = {
+			data: data,
+			headers: { 'Content-Type': 'application/json' },
+		}
+
+		if (extra_headers !== undefined) {
+			for (var header in extra_headers) {
+				args.headers[header] = extra_headers[header]
+			}
+		}
+
+		client
+			.put(url, args, function (data, response) {
+				cb(null, { data: data, response: response })
+			})
+			.on('error', function (error) {
+				debug('error response:', error)
+				cb(true, { error: error })
+			})
+	})
+
+	return self
 }
 
 exports = module.exports = function (system) {
-	return new rest(system);
-};
+	return new rest(system)
+}
