@@ -15,76 +15,74 @@
  *
  */
 
-var debug = require('debug')('lib/server_tcp');
-var net   = require('net');
+var debug = require('debug')('lib/server_tcp')
+var net = require('net')
 
 function server_tcp(system) {
+	var self = this
 
-	var self = this;
-
-	self.ready = true;
-	self.system = system;
-	self.clients = [];
+	self.ready = true
+	self.system = system
+	self.clients = []
 
 	self.server = net.createServer(function (socket) {
-
 		socket.on('end', function () {
-			self.clients.splice(self.clients.indexOf(socket), 1);
-			debug('Client disconnected: ' + socket.name);
-			system.emit('log', 'TCP Server', 'debug', 'Client disconnected: ' + socket.name);
-		});
+			self.clients.splice(self.clients.indexOf(socket), 1)
+			debug('Client disconnected: ' + socket.name)
+			system.emit('log', 'TCP Server', 'debug', 'Client disconnected: ' + socket.name)
+		})
 
 		socket.on('error', function () {
-			self.clients.splice(self.clients.indexOf(socket), 1);
-			debug('Client debug disconnected: ' + socket.name);
-			system.emit('log', 'TCP Server', 'error', 'Client errored/died: ' + socket.name);
-		});
+			self.clients.splice(self.clients.indexOf(socket), 1)
+			debug('Client debug disconnected: ' + socket.name)
+			system.emit('log', 'TCP Server', 'error', 'Client errored/died: ' + socket.name)
+		})
 
-		socket.name = socket.remoteAddress + ":" + socket.remotePort;
-		self.clients.push(socket);
-		debug('Client connected: ' + socket.name);
+		socket.name = socket.remoteAddress + ':' + socket.remotePort
+		self.clients.push(socket)
+		debug('Client connected: ' + socket.name)
 
-		system.emit('log', 'TCP Server', 'debug', 'Client connected: ' + socket.name);
+		system.emit('log', 'TCP Server', 'debug', 'Client connected: ' + socket.name)
 
 		// separate buffered stream into lines with responses
-		var receivebuffer = "";
+		var receivebuffer = ''
 
 		socket.on('data', function (chunk) {
-			var i = 0, line = '', offset = 0;
-			receivebuffer += chunk;
-			while ( (i = receivebuffer.indexOf('\n', offset)) !== -1) {
-				line = receivebuffer.substr(offset, i - offset);
-				offset = i + 1;
-				system.emit('server_api_command', line.toString().replace(/\r/, ""), function(err, res) {
+			var i = 0,
+				line = '',
+				offset = 0
+			receivebuffer += chunk
+			while ((i = receivebuffer.indexOf('\n', offset)) !== -1) {
+				line = receivebuffer.substr(offset, i - offset)
+				offset = i + 1
+				system.emit('server_api_command', line.toString().replace(/\r/, ''), function (err, res) {
 					if (err == null) {
-						debug("TCP command succeeded");
+						debug('TCP command succeeded')
+					} else {
+						debug('TCP command failed')
 					}
-					else {
-						debug("TCP command failed")
-					}
-					socket.write(res + "\n");
-				});
+					socket.write(res + '\n')
+				})
 			}
-			receivebuffer = receivebuffer.substr(offset);
-		});
-
-	});
+			receivebuffer = receivebuffer.substr(offset)
+		})
+	})
 
 	self.server.on('error', function (e) {
-		debug('TCP Server got error: ' + e);
+		debug('TCP Server got error: ' + e)
 	})
 
 	try {
-		debug("Trying: listen to tcp 51234");
-		self.server.listen(51234);
-		system.emit('log', 'TCP Server', 'info', 'Listening for TCP commands on port 51234');
-	} catch(e) {
-		system.emit('log', 'TCP Server', 'error', 'Couldn\'t bind to TCP port 51234');
+		debug('Trying: listen to tcp 51234')
+		self.server.listen(51234)
+		system.emit('log', 'TCP Server', 'info', 'Listening for TCP commands on port 51234')
+	} catch (e) {
+		system.emit('log', 'TCP Server', 'error', "Couldn't bind to TCP port 51234")
 	}
 
-	return self;
+	return self
 }
 
 exports = module.exports = function (system) {
-	return new server_tcp(system);
-};
+	return new server_tcp(system)
+}
