@@ -15,44 +15,40 @@
  *
  */
 
-var debug = require('debug')('lib/server_udp');
-var dgram = require('dgram');
-var server = dgram.createSocket('udp4');
+var debug = require('debug')('lib/server_udp')
+var dgram = require('dgram')
+var server = dgram.createSocket('udp4')
 
 function server_udp(system) {
+	var self = this
 
-	var self = this;
-
-	self.system = system;
+	self.system = system
 
 	server.on('listening', function () {
-		var address = server.address();
-		debug('UDP Server listening on ' + address.address + ":" + address.port);
-	});
+		var address = server.address()
+		debug('UDP Server listening on ' + address.address + ':' + address.port)
+	})
 
 	server.on('message', function (data, remote) {
-		debug(remote.address + ':' + remote.port +' received packet: ' + data.toString().trim() );
-		system.emit('server_api_command', data.toString(), function(err, res) {
+		debug(remote.address + ':' + remote.port + ' received packet: ' + data.toString().trim())
+		system.emit('server_api_command', data.toString(), function (err, res) {
 			if (err == null) {
-				debug("UDP command succeeded");
+				debug('UDP command succeeded')
+			} else {
+				debug('UDP command failed')
 			}
-			else {
-				debug("UDP command failed")
-			}
-		});
-
-	});
+		})
+	})
 
 	try {
-		debug("Trying: listen to tcp 51234");
-		system.emit('log', 'UDP Server', 'info', 'Listening for UDP commands on port 51235');
-		server.bind(51235, '0.0.0.0');
-	} catch(e) {
-		system.emit('log', 'UDP Server', 'error', 'Couldn\'t bind to UDP port 51235');
+		debug('Trying: listen to udp 51235')
+		system.emit('log', 'UDP Server', 'info', 'Listening for UDP commands on port 51235')
+		server.bind(51235, '0.0.0.0')
+	} catch (e) {
+		system.emit('log', 'UDP Server', 'error', "Couldn't bind to UDP port 51235")
 	}
-
 }
 
 exports = module.exports = function (system) {
-	return new server_udp(system);
-};
+	return new server_udp(system)
+}
