@@ -15,37 +15,31 @@
  *
  */
 
-var debug   = require('debug')('lib/tablet');
-var system;
+var debug = require('debug')('lib/tablet')
+var system
 
 function tablet(system) {
-	var self = this;
+	var self = this
 
-	self.tablet = {};
+	self.tablet = {}
 
-	system.on('get_tablet', function(cb) {
-		cb(self.tablet);
-	});
+	system.on('get_tablet', function (cb) {
+		cb(self.tablet)
+	})
 
 	system.emit('io_get', function (io) {
-
 		system.on('io_connect', function (socket) {
+			debug('socket ' + socket.id + ' connected')
 
-			debug('socket ' + socket.id + ' connected');
-
-			socket.on('tablet_startup', function() {
-
-			});
+			socket.on('tablet_startup', function () {})
 
 			socket.on('disconnect', function () {
-				debug('socket ' + socket.id + ' disconnected');
-			});
-
-		});
-	});
-
+				debug('socket ' + socket.id + ' disconnected')
+			})
+		})
+	})
 }
 
 module.exports = function (system) {
-	return new tablet(system);
-};
+	return new tablet(system)
+}
