@@ -15,53 +15,56 @@
  *
  */
 
-
-var util         = require('util');
-var debug        = require('debug')('lib/usb/elgato_mini');
-var elgato_base  = require('./elgato-base');
-var sharp        = require('sharp');
-var image_write_queue = require('./image-write-queue');
+var util = require('util')
+var debug = require('debug')('lib/usb/elgato_mini')
+var elgato_base = require('./elgato-base')
+var sharp = require('sharp')
+var image_write_queue = require('./image-write-queue')
 
 function elgato_mini(system, devicepath) {
-	var self = this;
+	var self = this
 
-	elgato_base.apply(this, [system, devicepath, 'Elgato Streamdeck Mini']);
+	elgato_base.apply(this, [system, devicepath, 'Elgato Streamdeck Mini'])
 
-	self.write_queue = new image_write_queue(function(key, buffer) {
+	self.write_queue = new image_write_queue(function (key, buffer) {
 		return sharp(buffer, { raw: { width: 72, height: 72, channels: 3 } })
 			.resize(80, 80)
 			.raw()
 			.toBuffer()
 			.catch(function (e) {
-				self.system.emit('log', 'device('+self.serialnumber+')', 'debug', `scale image failed: ${e}`);
+				self.system.emit('log', 'device(' + self.serialnumber + ')', 'debug', `scale image failed: ${e}`)
+				self.system.emit('elgatodm_remove_device', self.devicepath)
 			})
 			.then(function (newbuffer) {
-				self.streamDeck.fillImage(key, newbuffer);
-			}).catch(function (e) {
-				self.system.emit('log', 'device('+self.serialnumber+')', 'debug', `fillImage failed: ${e}`);
-				self.system.emit('elgatodm_remove_device', self.devicepath);
+				if (newbuffer) {
+					self.streamDeck.fillImage(key, newbuffer)
+				}
+			})
+			.catch(function (e) {
+				self.system.emit('log', 'device(' + self.serialnumber + ')', 'debug', `fillImage failed: ${e}`)
+				self.system.emit('elgatodm_remove_device', self.devicepath)
 			})
 	})
 
-	return self;
+	return self
 }
-elgato_mini.device_type = 'StreamDeck Mini';
+elgato_mini.device_type = 'StreamDeck Mini'
 
-elgato_mini.prototype.draw = function(key, buffer) {
-	var self = this;
+elgato_mini.prototype.draw = function (key, buffer) {
+	var self = this
 
-	var button = self.toDeviceKey(key);
+	var button = self.toDeviceKey(key)
 	if (button < 0 || button >= 6) {
-		return true;
+		return true
 	}
 
-	buffer = self.handleBuffer(buffer);
+	buffer = self.handleBuffer(buffer)
 
-	self.write_queue.queue(button, buffer);
+	self.write_queue.queue(button, buffer)
 
-	return true;
+	return true
 }
 
-util.inherits(elgato_mini, elgato_base);
+util.inherits(elgato_mini, elgato_base)
 
-exports = module.exports = elgato_mini;
+exports = module.exports = elgato_mini
