@@ -15,48 +15,51 @@
  *
  */
 
-
-var util         = require('util');
-var debug        = require('debug')('lib/usb/elgato_xl');
-var elgato_base  = require('./elgato-base');
-var sharp        = require('sharp');
-var image_write_queue = require('./image-write-queue');
+var util = require('util')
+var debug = require('debug')('lib/usb/elgato_xl')
+var elgato_base = require('./elgato-base')
+var sharp = require('sharp')
+var image_write_queue = require('./image-write-queue')
 
 function elgato_xl(system, devicepath) {
-	var self = this;
+	var self = this
 
-	elgato_base.apply(this, [system, devicepath, 'Elgato Streamdeck XL']);
+	elgato_base.apply(this, [system, devicepath, 'Elgato Streamdeck XL'])
 
-	self.write_queue = new image_write_queue(function(key, buffer) {
+	self.write_queue = new image_write_queue(function (key, buffer) {
 		return sharp(buffer, { raw: { width: 72, height: 72, channels: 3 } })
 			.resize(96, 96)
 			.raw()
 			.toBuffer()
 			.catch(function (e) {
-				self.system.emit('log', 'device('+self.serialnumber+')', 'debug', `scale image failed: ${e}`);
+				self.system.emit('log', 'device(' + self.serialnumber + ')', 'debug', `scale image failed: ${e}`)
+				self.system.emit('elgatodm_remove_device', self.devicepath)
 			})
 			.then(function (newbuffer) {
-				self.streamDeck.fillImage(key, newbuffer);
-			}).catch(function (e) {
-				self.system.emit('log', 'device('+self.serialnumber+')', 'debug', `fillImage failed: ${e}`);
-				self.system.emit('elgatodm_remove_device', self.devicepath);
+				if (newbuffer) {
+					self.streamDeck.fillImage(key, newbuffer)
+				}
+			})
+			.catch(function (e) {
+				self.system.emit('log', 'device(' + self.serialnumber + ')', 'debug', `fillImage failed: ${e}`)
+				self.system.emit('elgatodm_remove_device', self.devicepath)
 			})
 	})
 
-	return self;
+	return self
 }
-elgato_xl.device_type = 'StreamDeck XL';
+elgato_xl.device_type = 'StreamDeck XL'
 
-elgato_xl.prototype.draw = function(key, buffer, ts) {
-	var self = this;
+elgato_xl.prototype.draw = function (key, buffer, ts) {
+	var self = this
 
-	buffer = self.handleBuffer(buffer);
+	buffer = self.handleBuffer(buffer)
 
-	self.write_queue.queue(key, buffer);
+	self.write_queue.queue(key, buffer)
 
-	return true;
+	return true
 }
 
-util.inherits(elgato_xl, elgato_base);
+util.inherits(elgato_xl, elgato_base)
 
-exports = module.exports = elgato_xl;
+exports = module.exports = elgato_xl
