@@ -15,49 +15,47 @@
  *
  */
 
-
-var util         = require('util');
-var debug        = require('debug')('lib/usb/elgato');
-var elgato_base  = require('./elgato-base');
+var util = require('util')
+var debug = require('debug')('lib/usb/elgato')
+var elgato_base = require('./elgato-base')
 
 function elgato(system, devicepath) {
-	var self = this;
+	var self = this
 
-	elgato_base.apply(this, [system, devicepath, 'Elgato Streamdeck']);
+	elgato_base.apply(this, [system, devicepath, 'Elgato Streamdeck'])
 
-	return self;
+	return self
 }
-elgato.device_type = 'StreamDeck';
+elgato.device_type = 'StreamDeck'
 
-elgato.prototype.draw = function(key, buffer, attempts) {
-	var self = this;
+elgato.prototype.draw = function (key, buffer, attempts) {
+	var self = this
 
 	// null/undefined => 0
-	attempts = ~~attempts;
+	attempts = ~~attempts
 
 	if (attempts === 0) {
-		buffer = self.handleBuffer(buffer);
+		buffer = self.handleBuffer(buffer)
 	}
 
-	attempts++;
+	attempts++
 
-	var drawKey = self.toDeviceKey(key);
+	var drawKey = self.toDeviceKey(key)
 
 	try {
-
 		if (drawKey !== undefined && drawKey >= 0 && drawKey < self.info.keysTotal) {
-			self.streamDeck.fillImage(drawKey, buffer);
+			self.streamDeck.fillImage(drawKey, buffer)
 		}
 
-		return true;
+		return true
 	} catch (e) {
-		self.log('StreamDeck USB Exception: ' + e.message);
+		self.log('StreamDeck USB Exception: ' + e.message)
 
 		if (attempts > 2) {
-			self.log('Giving up USB device ' + self.devicepath);
-			self.system.emit('elgatodm_remove_device', self.devicepath);
+			self.log('Giving up USB device ' + self.devicepath)
+			self.system.emit('elgatodm_remove_device', self.devicepath)
 
-			return false;
+			return false
 		}
 
 		setTimeout(self.draw.bind(self), 20, key, buffer, attempts)
@@ -65,6 +63,6 @@ elgato.prototype.draw = function(key, buffer, attempts) {
 	}
 }
 
-util.inherits(elgato, elgato_base);
+util.inherits(elgato, elgato_base)
 
-exports = module.exports = elgato;
+exports = module.exports = elgato
