@@ -15,73 +15,71 @@
  *
  */
 
- /*
+/*
 	This is written with the idea that it might only handle one device, or more than one device
 	for future compatibility. Currently each handler instance only handles one device though.
 */
 
 var drivers = {
-	'elgato': require('./elgato'),
+	elgato: require('./elgato'),
 	'elgato-v2': require('./elgato'),
 	'elgato-mini': require('./elgato-mini'),
 	'elgato-xl': require('./elgato-xl'),
-	'infinitton': require('./infinitton')
-};
+	infinitton: require('./infinitton'),
+	xkeys: require('./xkeys'),
+}
 
-global.MAX_BUTTONS = parseInt(process.env.MAX_BUTTONS);
-global.MAX_BUTTONS_PER_ROW = parseInt(process.env.MAX_BUTTONS_PER_ROW);
+global.MAX_BUTTONS = parseInt(process.env.MAX_BUTTONS)
+global.MAX_BUTTONS_PER_ROW = parseInt(process.env.MAX_BUTTONS_PER_ROW)
 
-var EventEmitter = require('events').EventEmitter;
-var util = require('util');
-var driver_instances = {};
+var EventEmitter = require('events').EventEmitter
+var util = require('util')
+var driver_instances = {}
 
 function eventhandler() {
-	EventEmitter.call(this);
-};
-util.inherits(eventhandler, EventEmitter);
+	EventEmitter.call(this)
+}
+util.inherits(eventhandler, EventEmitter)
 
 eventhandler.prototype.emit = function () {
-	var args = [].slice.call(arguments);
-	process.send({ cmd: 'system', args: args });
-	EventEmitter.prototype.emit.apply(this, args);
-};
+	var args = [].slice.call(arguments)
+	process.send({ cmd: 'system', args: args })
+	EventEmitter.prototype.emit.apply(this, args)
+}
 
-var system = new eventhandler();
+var system = new eventhandler()
 
 process.on('message', function (data) {
 	if (data.cmd == 'add') {
 		try {
-			var type = drivers[data.type];
-			driver_instances[data.id] = new type(system, data.devicepath);
+			var type = drivers[data.type]
+			driver_instances[data.id] = new type(system, data.devicepath)
 			driver_instances[data.id].log = function () {
-				var args = [].slice.call(arguments);
-				process.send({ cmd: 'debug', id: data.id, args: args });
-			};
-			process.send({ cmd: 'publish', id: data.id, info: driver_instances[data.id].info });
-			process.send({ cmd: 'add', id: data.id });
+				var args = [].slice.call(arguments)
+				process.send({ cmd: 'debug', id: data.id, args: args })
+			}
+			process.send({ cmd: 'publish', id: data.id, info: driver_instances[data.id].info })
+			process.send({ cmd: 'add', id: data.id })
 		} catch (e) {
-			process.send({ cmd: 'error', id: data.id, error: e.message + ' AND GOT: ' + JSON.stringify(data) });
+			process.send({ cmd: 'error', id: data.id, error: e.message + ' AND GOT: ' + JSON.stringify(data) })
 		}
-	}
-	else if (data.cmd == 'execute') {
+	} else if (data.cmd == 'execute') {
 		try {
-			var result = driver_instances[data.id][data.function].apply(driver_instances[data.id], data.args);
+			var result = driver_instances[data.id][data.function].apply(driver_instances[data.id], data.args)
 			if (data.returnId !== undefined) {
-				process.send({ cmd: 'return', id: data.id, returnId: data.returnId, result: result });
+				process.send({ cmd: 'return', id: data.id, returnId: data.returnId, result: result })
 			}
 		} catch (e) {
-			process.send({ cmd: 'error', id: data.id, error: e.message });
+			process.send({ cmd: 'error', id: data.id, error: e.message })
 		}
-	}
-	else if (data.cmd == 'system') {
-		EventEmitter.prototype.emit.apply(system, args);
-	}
-	else if (data.cmd == 'remove') {
+	} else if (data.cmd == 'system') {
+		EventEmitter.prototype.emit.apply(system, data.args)
+	} else if (data.cmd == 'remove') {
 		try {
-			driver_instances[data.id].quit();
-			process.send({ cmd: 'remove', id: data.id });
+			driver_instances[data.id].quit()
+			process.send({ cmd: 'remove', id: data.id })
 		} catch (e) {
-			process.send({ cmd: 'error', id: data.id, error: e.message });
+			process.send({ cmd: 'error', id: data.id, error: e.message })
 		}
 	}
-});
+})
