@@ -15,19 +15,19 @@
  *
  */
 
-var debug = require('debug')('lib/usb/image_write_queue');
+var debug = require('debug')('lib/usb/image_write_queue')
 
 function image_write_queue(fillImage) {
-	var self = this;
+	var self = this
 
 	self.maxConcurrent = 3
-	
+
 	self.fillImage = fillImage
 	self.pendingImages = []
 	self.inProgress = []
 }
-image_write_queue.prototype.queue = function(key, buffer) {
-	var self = this;
+image_write_queue.prototype.queue = function (key, buffer) {
+	var self = this
 
 	let updated = false
 	// Try and replace an existing queued image first
@@ -46,13 +46,13 @@ image_write_queue.prototype.queue = function(key, buffer) {
 
 	self.tryDequeue()
 }
-image_write_queue.prototype.tryDequeue = function() {
-	var self = this;
+image_write_queue.prototype.tryDequeue = function () {
+	var self = this
 
 	// Start another if not too many in progress
 	if (self.inProgress.length <= self.maxConcurrent && self.pendingImages.length > 0) {
 		// Find first image where key is not being worked on
-		const nextImageIndex = self.pendingImages.findIndex(img => self.inProgress.indexOf(img.key) === -1)
+		const nextImageIndex = self.pendingImages.findIndex((img) => self.inProgress.indexOf(img.key) === -1)
 		if (nextImageIndex === -1) {
 			return
 		}
@@ -66,21 +66,22 @@ image_write_queue.prototype.tryDequeue = function() {
 		// Track which key is being processed
 		self.inProgress.push(nextImage.key)
 
-		self.fillImage(nextImage.key, nextImage.buffer)
-		.catch(e => {
-            // Ensure it doesnt error out
-            debug('fillImage error:', e)
-        })
-		.then(() => {
-			// Stop tracking key
-			self.inProgress = self.inProgress.filter(k => k !== nextImage.key)
-
-			// Run again
-			setImmediate(() => {
-				self.tryDequeue()
+		self
+			.fillImage(nextImage.key, nextImage.buffer)
+			.catch((e) => {
+				// Ensure it doesnt error out
+				debug('fillImage error:', e)
 			})
-		})
+			.then(() => {
+				// Stop tracking key
+				self.inProgress = self.inProgress.filter((k) => k !== nextImage.key)
+
+				// Run again
+				setImmediate(() => {
+					self.tryDequeue()
+				})
+			})
 	}
 }
 
-exports = module.exports = image_write_queue;
+exports = module.exports = image_write_queue
