@@ -15,58 +15,62 @@
  *
  */
 
-var debug   = require('debug')('lib/userconfig');
-var system;
+var debug = require('debug')('lib/userconfig')
+var system
 
 function userconfig(system) {
-	var self = this;
+	var self = this
 
-	self.userconfig = {};
+	self.userconfig = {}
 
-	system.emit('db_get', 'userconfig', function(config) {
+	system.emit('db_get', 'userconfig', function (config) {
 		if (config === undefined) {
-			config = {};
+			config = {}
 		}
 
-		self.userconfig = config;
+		self.userconfig = config
 
 		for (var key in config) {
-			system.emit('set_userconfig_key', key, config[key]);
+			system.emit('set_userconfig_key', key, config[key])
 		}
-	});
+	})
 
-	system.on('get_userconfig', function(cb) {
-		cb(self.userconfig);
-	});
+	system.on('get_userconfig', function (cb) {
+		cb(self.userconfig)
+	})
 
 	system.emit('io_get', function (io) {
-
 		system.on('io_connect', function (socket) {
+			function sendResult(answer, name, ...args) {
+				if (typeof answer === 'function') {
+					answer(...args)
+				} else {
+					socket.emit(name, ...args)
+				}
+			}
 
-			debug('socket ' + socket.id + ' connected');
+			debug('socket ' + socket.id + ' connected')
 
-			socket.on('set_userconfig_key', function(key,value) {
-				self.userconfig[key] = value;
-				debug('set_userconfig_key', key, value);
-				system.emit('log', 'set_userconfig('+key+')', 'info', 'new value: ' + value);
-				socket.broadcast.emit('set_userconfig_key', key, value);
-				system.emit('set_userconfig_key', key, value);
-				system.emit('db_save');
-			});
+			socket.on('set_userconfig_key', function (key, value) {
+				self.userconfig[key] = value
+				debug('set_userconfig_key', key, value)
+				system.emit('log', 'set_userconfig(' + key + ')', 'info', 'new value: ' + value)
+				io.emit('set_userconfig_key', key, value)
+				system.emit('set_userconfig_key', key, value)
+				system.emit('db_save')
+			})
 
-			socket.on('get_userconfig_all', function() {
-				socket.emit('get_userconfig_all', self.userconfig);
-			});
+			socket.on('get_userconfig_all', function (answer) {
+				sendResult(answer, 'get_userconfig_all', self.userconfig)
+			})
 
 			socket.on('disconnect', function () {
-				debug('socket ' + socket.id + ' disconnected');
-			});
-
-		});
-	});
-
+				debug('socket ' + socket.id + ' disconnected')
+			})
+		})
+	})
 }
 
 module.exports = function (system) {
-	return new userconfig(system);
-};
+	return new userconfig(system)
+}
