@@ -15,195 +15,211 @@
  *
  */
 
-var system;
-var io;
-var debug = require('debug')('lib/variable');
+var system
+var io
+var debug = require('debug')('lib/variable')
 
 function variable(system) {
-	var self = this;
+	var self = this
 
-	self.system = system;
-	self.variable_definitions = {};
-	self.variables = {};
+	self.system = system
+	self.variable_definitions = {}
+	self.variables = {}
 
 	system.emit('io_get', function (_io) {
-		io = _io;
+		io = _io
 
 		system.on('io_connect', function (socket) {
-			socket.on('variable_instance_definitions_get', function () {
-				socket.emit('variable_instance_definitions_get:result', null, self.variable_definitions);
-			});
+			function sendResult(answer, name, ...args) {
+				if (typeof answer === 'function') {
+					answer(...args)
+				} else {
+					socket.emit(name, ...args)
+				}
+			}
 
-			socket.on('variables_get', function () {
-				var vars = {};
+			socket.on('variable_instance_definitions_get', function (answer) {
+				sendResult(answer, 'variable_instance_definitions_get:result', self.variable_definitions)
+			})
+
+			socket.on('variables_get', function (answer) {
+				var vars = {}
 				for (var label in self.variables) {
 					for (var variable in self.variables[label]) {
-						vars[label + ':' + variable] = self.variables[label][variable];
+						vars[label + ':' + variable] = self.variables[label][variable]
 					}
 				}
 
-				socket.emit('variables_get:result', null, vars);
-			});
-		});
-	});
+				sendResult(answer, 'variables_get:result', vars)
+			})
+		})
+	})
 
 	system.on('variable_get_definitions', function (cb) {
-		cb(self.variable_definitions);
-	});
+		cb(self.variable_definitions)
+	})
 
 	system.on('variable_instance_definitions_set', function (instance, variables) {
-		self.variable_definitions[instance.label] = variables;
+		self.variable_definitions[instance.label] = variables
 
-		debug('got instance variable definitions for ' + instance.label);
-		io.emit('variable_instance_definitions_set', instance.label, variables);
-	});
+		debug('got instance variable definitions for ' + instance.label)
+		io.emit('variable_instance_definitions_set', instance.label, variables)
+	})
 
 	system.on('variable_instance_set', function (instance, variable, value) {
-		self.set_variable(instance.label, variable, value);
-	});
+		self.set_variable(instance.label, variable, value)
+	})
 
 	system.on('variable_rename_callback', function (str, fromlabel, tolabel, cb) {
-			if (typeof str != 'string') {
-				console.log("Warning, variable_rename_callback was called with this: ", str);
-				return cb(str);
-			}
-			var fixtext = str;
+		if (typeof str != 'string') {
+			console.log('Warning, variable_rename_callback was called with this: ', str)
+			return cb(str)
+		}
+		var fixtext = str
 
-			if (fixtext.match(/\$\(/)) {
-				var matches, reg = /\$\(([^:)]+):([^)]+)\)/g;
+		if (fixtext.match(/\$\(/)) {
+			var matches,
+				reg = /\$\(([^:)]+):([^)]+)\)/g
 
-				while ((matches = reg.exec(fixtext)) !== null) {
-					if (matches[1] !== undefined && matches[1] == fromlabel) {
-						if (matches[2] !== undefined) {
-							reg2 = new RegExp('\\$\\(' + escape(matches[1]) + ':' + escape(matches[2]) + '\\)');
-							str = str.replace(reg2, '$(' + tolabel + ':' + matches[2] + ')');
-						}
+			while ((matches = reg.exec(fixtext)) !== null) {
+				if (matches[1] !== undefined && matches[1] == fromlabel) {
+					if (matches[2] !== undefined) {
+						reg2 = new RegExp('\\$\\(' + escape(matches[1]) + ':' + escape(matches[2]) + '\\)')
+						str = str.replace(reg2, '$(' + tolabel + ':' + matches[2] + ')')
 					}
 				}
 			}
+		}
 
-			cb(str);
-	});
+		cb(str)
+	})
 
 	system.on('variable_instance_label_rename', function (labelFrom, labelTo) {
 		if (self.variables[labelTo] === undefined) {
-			self.variables[labelTo] = {};
+			self.variables[labelTo] = {}
 		}
 		if (self.variables[labelFrom] !== undefined) {
 			for (var variable in self.variables[labelFrom]) {
-				system.emit('bank_rename_variables', labelFrom, labelTo);
-				self.variables[labelTo][variable] = self.variables[labelFrom][variable];
-				delete self.variables[labelFrom][variable];
-				io.emit('variable_set', labelFrom + ':' + variable, undefined);
+				system.emit('bank_rename_variables', labelFrom, labelTo)
+				self.variables[labelTo][variable] = self.variables[labelFrom][variable]
+				delete self.variables[labelFrom][variable]
+				io.emit('variable_set', labelFrom + ':' + variable, undefined)
 
 				// In case variables exists in banks from before
-				self.system.emit('variable_changed', labelTo, variable, self.variables[labelTo][variable]);
-				io.emit('variable_set', labelTo + ':' + variable, self.variables[labelTo][variable]);
+				self.system.emit('variable_changed', labelTo, variable, self.variables[labelTo][variable])
+				io.emit('variable_set', labelTo + ':' + variable, self.variables[labelTo][variable])
 			}
-			delete self.variables[labelFrom];
+			delete self.variables[labelFrom]
 		}
 
 		if (self.variable_definitions[labelFrom] !== undefined) {
-			self.variable_definitions[labelTo] = self.variable_definitions[labelFrom];
-			delete self.variable_definitions[labelFrom];
+			self.variable_definitions[labelTo] = self.variable_definitions[labelFrom]
+			delete self.variable_definitions[labelFrom]
 
-			io.emit('variable_instance_definitions_set', labelTo, self.variable_definitions[labelTo]);
-			io.emit('variable_instance_definitions_set', labelFrom, []);
+			io.emit('variable_instance_definitions_set', labelTo, self.variable_definitions[labelTo])
+			io.emit('variable_instance_definitions_set', labelFrom, [])
 		}
-	});
+	})
 
 	system.on('instance_enable', function (id, state) {
 		if (state === false) {
 			system.emit('instance_get', id, function (info) {
 				if (info && self.variables[info.label] !== undefined) {
-					var keys = Object.keys(self.variables[info.label]);
-					delete self.variable_definitions[info.label];
-					delete self.variables[info.label];
-					io.emit('variable_instance_definitions_set', info.label, []);
+					var keys = Object.keys(self.variables[info.label])
+					delete self.variable_definitions[info.label]
+					delete self.variables[info.label]
+					io.emit('variable_instance_definitions_set', info.label, [])
 
 					// Reset banks
 					for (var i = 0; i < keys.length; ++i) {
 						// Force update
-						self.system.emit('variable_changed', info.label, keys[i], undefined);
+						self.system.emit('variable_changed', info.label, keys[i], undefined)
 					}
 				}
-			});
+			})
 		}
-	});
+	})
 
 	system.on('instance_delete', function (id, label) {
-		io.emit('variable_instance_definitions_set', label, []);
+		io.emit('variable_instance_definitions_set', label, [])
 
 		if (label !== undefined) {
 			if (self.variables[label] !== undefined) {
 				for (var variable in self.variables[label]) {
-					self.variables[label][variable] = undefined;
-					self.system.emit('variable_changed', label, variable, undefined);
-					io.emit('variable_set', label + ':' + variable, undefined);
+					self.variables[label][variable] = undefined
+					self.system.emit('variable_changed', label, variable, undefined)
+					io.emit('variable_set', label + ':' + variable, undefined)
 				}
 			}
 
-			delete self.variable_definitions[label];
-			delete self.variables[label];
+			delete self.variable_definitions[label]
+			delete self.variables[label]
 
-			io.emit('variable_instance_definitions_set', label, []);
+			io.emit('variable_instance_definitions_set', label, [])
 		}
-	});
+	})
 
 	function escapereg(str) {
-		return str.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&');
+		return str.replace(/[-\/\\^$*+?.()|[\]{}]/g, '\\$&')
 	}
 
 	// Everybody stand back. I know regular expressions. - xckd #208 /ck/kc/
 	system.on('variable_parse', function (string, cb) {
-		var vars, matches, reg = /\$\([^:$)]+:[^)$]+\)/;
+		var vars,
+			matches,
+			reg = /\$\([^:$)]+:[^)$]+\)/
 
 		if (string === undefined) {
-			return cb(undefined);
+			return cb(undefined)
 		}
 
-		while (vars = reg.exec(string)) {
-			matches = vars[0].match(/\$\(([^:)]+):([^)]+)\)/);
+		while ((vars = reg.exec(string))) {
+			matches = vars[0].match(/\$\(([^:)]+):([^)]+)\)/)
 			if (self.variables[matches[1]] !== undefined) {
 				if (self.variables[matches[1]][matches[2]] !== undefined) {
-					string = string.replace(new RegExp(escapereg(matches[0])), self.variables[matches[1]][matches[2]]);
+					string = string.replace(new RegExp(escapereg(matches[0])), self.variables[matches[1]][matches[2]])
 				} else {
-					string = string.replace(new RegExp(escapereg(matches[0])), '$NA');
+					string = string.replace(new RegExp(escapereg(matches[0])), '$NA')
 				}
 			} else {
-				string = string.replace(new RegExp(escapereg(matches[0])), '$NA');
+				string = string.replace(new RegExp(escapereg(matches[0])), '$NA')
 			}
 		}
-		cb(string);
-	});
+		cb(string)
+	})
 
 	system.on('variable_get', function (label, variable, cb) {
 		if (self.variables[label] !== undefined) {
-			cb(self.variables[label][variable]);
+			cb(self.variables[label][variable])
 		} else {
-			cb(undefined);
+			cb(undefined)
 		}
-	});
+	})
 
-	return self;
+	return self
 }
 
-variable.prototype.set_variable = function(label, variable, value) {
-	var self = this;
+variable.prototype.set_variable = function (label, variable, value) {
+	var self = this
 
 	if (self.variables[label] === undefined) {
-		self.variables[label] = {};
+		self.variables[label] = {}
 	}
 
 	if (self.variables[label][variable] != value) {
-		self.variables[label][variable] = value;
+		self.variables[label][variable] = value
 
-		self.system.emit('variable_changed', label, variable, value);
-		debug('Variable $(' + label + ':' + variable + ') is "' + value + '"');
-		io.emit('variable_set', label + ':' + variable, value);
+		self.system.emit('variable_changed', label, variable, value)
+
+		// Skip debug if it's just internal:time_* spamming.
+		if (!(label === 'internal' && variable.match(/^time_/))) {
+			debug('Variable $(' + label + ':' + variable + ') is "' + value + '"')
+		}
+
+		io.emit('variable_set', label + ':' + variable, value)
 	}
-};
+}
 
 exports = module.exports = function (system) {
-	return new variable(system);
-};
+	return new variable(system)
+}
