@@ -15,49 +15,48 @@
  *
  */
 
- 
-var request = require("request")
+var request = require('request')
 
-var get = function(url,cb) {
+var get = function (url, cb) {
 	var options = {
 		url: url,
-		 headers: {
-			'User-Agent': 'request'
+		headers: {
+			'User-Agent': 'request',
 		},
-		json: true
-	};
+		json: true,
+	}
 
 	request(options, function (error, response, body) {
-
 		if (!error) {
-			cb(null, response, body);
-			return;
+			cb(null, response, body)
+			return
 		}
 
-		cb(error, null, null);
-
-	});
-};
+		cb(error, null, null)
+	})
+}
 
 var lineReader = require('readline').createInterface({
-	input: require('fs').createReadStream('.gitmodules')
-});
+	input: require('fs').createReadStream('.gitmodules'),
+})
 
 lineReader.on('line', function (line) {
-	var found;
-	if (found = line.match('companion-module-(.+).git')) {
-		(function(module) {
-			get("http://api.github.com/repos/bitfocus/companion-module-"+module+"/issues", function(err, response, body) {
-				if (err === null) {
-					for (var i in body) {
-						var issue = body[i];
-						console.log("["+module+"] #" + issue.number+ ": " + issue.title);
+	var found
+	if ((found = line.match('companion-module-(.+).git'))) {
+		;(function (module) {
+			get(
+				'http://api.github.com/repos/bitfocus/companion-module-' + module + '/issues',
+				function (err, response, body) {
+					if (err === null) {
+						for (var i in body) {
+							var issue = body[i]
+							console.log('[' + module + '] #' + issue.number + ': ' + issue.title)
+						}
+					} else {
+						console.log('error:', err)
 					}
 				}
-				else {
-					console.log("error:",err);
-				}
-			});
+			)
 		})(found[1])
 	}
-});
+})
